Network.AWS.Route53Domains.CheckDomainAvailability

Description

This operation checks the availability of one domain name. You can access this API without authenticating. Note that if the availability status of a domain is pending, you must submit another request to determine the availability of the domain name.

http://docs.aws.amazon.com/Route53/latest/APIReference/api-CheckDomainAvailability.html

Request

data CheckDomainAvailability

Request constructor

checkDomainAvailability

Arguments

:: Text
cdaDomainName

-> CheckDomainAvailability

CheckDomainAvailability constructor.

The fields accessible through corresponding lenses are:

cdaDomainName :: Text

cdaIdnLangCode :: Maybe Text

Request lenses

cdaDomainName :: Lens' CheckDomainAvailability Text

The name of a domain.

Type: String

Default: None

Constraints: The domain name can contain only the letters a through z, the numbers 0 through 9, and hyphen (-). Internationalized Domain Names are not supported.

Required: Yes

cdaIdnLangCode :: Lens' CheckDomainAvailability (Maybe Text)

Reserved for future use.

Response

data CheckDomainAvailabilityResponse

Response constructor

checkDomainAvailabilityResponse

Arguments

:: DomainAvailability
cdarAvailability

-> CheckDomainAvailabilityResponse

CheckDomainAvailabilityResponse constructor.

The fields accessible through corresponding lenses are:

cdarAvailability :: DomainAvailability

Response lenses

cdarAvailability :: Lens' CheckDomainAvailabilityResponse DomainAvailability

Whether the domain name is available for registering.

You can only register domains designated as AVAILABLE.

Type: String

Valid values:

AVAILABLE – The domain name is available. AVAILABLE_RESERVED – The domain name is reserved under specific conditions. AVAILABLE_PREORDER – The domain name is available and can be preordered. UNAVAILABLE – The domain name is not available. UNAVAILABLE_PREMIUM – The domain name is not available. UNAVAILABLE_RESTRICTED – The domain name is forbidden. RESERVED – The domain name has been reserved for another person or organization.