Network.AWS.Route53Domains.DisableDomainTransferLock

Description

This operation removes the transfer lock on the domain (specifically the clientTransferProhibited status) to allow domain transfers. We recommend you refrain from performing this action unless you intend to transfer the domain to a different registrar. Successful submission returns an operation ID that you can use to track the progress and completion of the action. If the request is not completed successfully, the domain registrant will be notified by email.

Creating a Request

disableDomainTransferLock

Arguments

:: Text
ddtlDomainName

-> DisableDomainTransferLock

Creates a value of DisableDomainTransferLock with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddtlDomainName - The name of the domain that you want to remove the transfer lock for.

data DisableDomainTransferLock

The DisableDomainTransferLock request includes the following element.

See: disableDomainTransferLock smart constructor.

Request Lenses

ddtlDomainName :: Lens' DisableDomainTransferLock Text

The name of the domain that you want to remove the transfer lock for.

Destructuring the Response

disableDomainTransferLockResponse

Arguments

:: Int
ddtlrsResponseStatus

-> Text
ddtlrsOperationId

-> DisableDomainTransferLockResponse

Creates a value of DisableDomainTransferLockResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddtlrsResponseStatus - -- | The response status code.

ddtlrsOperationId - Identifier for tracking the progress of the request. To use this ID to query the operation status, use GetOperationDetail .

data DisableDomainTransferLockResponse

The DisableDomainTransferLock response includes the following element.

See: disableDomainTransferLockResponse smart constructor.

Response Lenses

ddtlrsResponseStatus :: Lens' DisableDomainTransferLockResponse Int

- | The response status code.

ddtlrsOperationId :: Lens' DisableDomainTransferLockResponse Text

Identifier for tracking the progress of the request. To use this ID to query the operation status, use GetOperationDetail .