Network.AWS.Route53Domains.EnableDomainAutoRenew

Description

This operation configures Amazon Route 53 to automatically renew the specified domain before the domain registration expires. The cost of renewing your domain registration is billed to your AWS account.

The period during which you can renew a domain name varies by TLD. For a list of TLDs and their renewal policies, see "Renewal, restoration, anddeletion times" on the website for our registrar partner, Gandi. Route 53 requires that you renew before the end of the renewal period that is listed on the Gandi website so we can complete processing before the deadline.

http://docs.aws.amazon.com/Route53/latest/APIReference/api-EnableDomainAutoRenew.html

Request

data EnableDomainAutoRenew

Request constructor

enableDomainAutoRenew

Arguments

:: Text
edarDomainName

-> EnableDomainAutoRenew

EnableDomainAutoRenew constructor.

The fields accessible through corresponding lenses are:

edarDomainName :: Text

Request lenses

edarDomainName :: Lens' EnableDomainAutoRenew Text

Response

data EnableDomainAutoRenewResponse

Response constructor

enableDomainAutoRenewResponse :: EnableDomainAutoRenewResponse

EnableDomainAutoRenewResponse constructor.