Network.AWS.Route53Domains.EnableDomainTransferLock

Description

This operation sets the transfer lock on the domain (specifically the clientTransferProhibited status) to prevent domain transfers. Successful submission returns an operation ID that you can use to track the progress and completion of the action. If the request is not completed successfully, the domain registrant will be notified by email.

http://docs.aws.amazon.com/Route53/latest/APIReference/api-EnableDomainTransferLock.html

Request

data EnableDomainTransferLock

Request constructor

enableDomainTransferLock

Arguments

:: Text
edtlDomainName

-> EnableDomainTransferLock

EnableDomainTransferLock constructor.

The fields accessible through corresponding lenses are:

edtlDomainName :: Text

Request lenses

edtlDomainName :: Lens' EnableDomainTransferLock Text

The name of a domain.

Type: String

Default: None

Constraints: The domain name can contain only the letters a through z, the numbers 0 through 9, and hyphen (-). Internationalized Domain Names are not supported.

Required: Yes

Response

data EnableDomainTransferLockResponse

Response constructor

enableDomainTransferLockResponse

Arguments

:: Text
edtlrOperationId

-> EnableDomainTransferLockResponse

EnableDomainTransferLockResponse constructor.

The fields accessible through corresponding lenses are:

edtlrOperationId :: Text

Response lenses

edtlrOperationId :: Lens' EnableDomainTransferLockResponse Text

Identifier for tracking the progress of the request. To use this ID to query the operation status, use GetOperationDetail.

Type: String

Default: None

Constraints: Maximum 255 characters.