Network.AWS.Route53Domains.GetContactReachabilityStatus

Description

For operations that require confirmation that the email address for the registrant contact is valid, such as registering a new domain, this operation returns information about whether the registrant contact has responded.

If you want us to resend the email, use the ResendContactReachabilityEmail operation.

Creating a Request

getContactReachabilityStatus :: GetContactReachabilityStatus

Creates a value of GetContactReachabilityStatus with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gcrsDomainName - The name of the domain for which you want to know whether the registrant contact has confirmed that the email address is valid.

data GetContactReachabilityStatus

See: getContactReachabilityStatus smart constructor.

Request Lenses

gcrsDomainName :: Lens' GetContactReachabilityStatus (Maybe Text)

The name of the domain for which you want to know whether the registrant contact has confirmed that the email address is valid.

Destructuring the Response

getContactReachabilityStatusResponse

Arguments

:: Int
gcrsrsResponseStatus

-> GetContactReachabilityStatusResponse

Creates a value of GetContactReachabilityStatusResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gcrsrsStatus - Whether the registrant contact has responded. Values include the following: * PENDING * We sent the confirmation email and haven't received a response yet. * DONE * We sent the email and got confirmation from the registrant contact. * EXPIRED * The time limit expired before the registrant contact responded.

gcrsrsDomainName - The domain name for which you requested the reachability status.

gcrsrsResponseStatus - -- | The response status code.

data GetContactReachabilityStatusResponse

See: getContactReachabilityStatusResponse smart constructor.

Response Lenses

gcrsrsStatus :: Lens' GetContactReachabilityStatusResponse (Maybe ReachabilityStatus)

Whether the registrant contact has responded. Values include the following: * PENDING * We sent the confirmation email and haven't received a response yet. * DONE * We sent the email and got confirmation from the registrant contact. * EXPIRED * The time limit expired before the registrant contact responded.

gcrsrsDomainName :: Lens' GetContactReachabilityStatusResponse (Maybe Text)

The domain name for which you requested the reachability status.

gcrsrsResponseStatus :: Lens' GetContactReachabilityStatusResponse Int

- | The response status code.