Network.AWS.Route53Domains.GetDomainDetail

Description

This operation returns detailed information about the domain. The domain's contact information is also returned as part of the output.

http://docs.aws.amazon.com/Route53/latest/APIReference/api-GetDomainDetail.html

Request

data GetDomainDetail

Request constructor

getDomainDetail

Arguments

:: Text
gddDomainName

-> GetDomainDetail

GetDomainDetail constructor.

The fields accessible through corresponding lenses are:

gddDomainName :: Text

Request lenses

gddDomainName :: Lens' GetDomainDetail Text

The name of a domain.

Type: String

Default: None

Constraints: The domain name can contain only the letters a through z, the numbers 0 through 9, and hyphen (-). Internationalized Domain Names are not supported.

Required: Yes

Response

data GetDomainDetailResponse

Response constructor

getDomainDetailResponse

Arguments

:: Text
gddrDomainName

-> ContactDetail
gddrAdminContact

-> ContactDetail
gddrRegistrantContact

-> ContactDetail
gddrTechContact

-> GetDomainDetailResponse

GetDomainDetailResponse constructor.

The fields accessible through corresponding lenses are:

gddrAbuseContactEmail :: Maybe Text

gddrAbuseContactPhone :: Maybe Text

gddrAdminContact :: ContactDetail

gddrAdminPrivacy :: Maybe Bool

gddrAutoRenew :: Maybe Bool

gddrCreationDate :: Maybe UTCTime

gddrDnsSec :: Maybe Text

gddrDomainName :: Text

gddrExpirationDate :: Maybe UTCTime

gddrNameservers :: [Nameserver]

gddrRegistrantContact :: ContactDetail

gddrRegistrantPrivacy :: Maybe Bool

gddrRegistrarName :: Maybe Text

gddrRegistrarUrl :: Maybe Text

gddrRegistryDomainId :: Maybe Text

gddrReseller :: Maybe Text

gddrStatusList :: [Text]

gddrTechContact :: ContactDetail

gddrTechPrivacy :: Maybe Bool

gddrUpdatedDate :: Maybe UTCTime

gddrWhoIsServer :: Maybe Text

Response lenses

gddrAbuseContactEmail :: Lens' GetDomainDetailResponse (Maybe Text)

Email address to contact to report incorrect contact information for a domain, to report that the domain is being used to send spam, to report that someone is cybersquatting on a domain name, or report some other type of abuse.

Type: String

gddrAbuseContactPhone :: Lens' GetDomainDetailResponse (Maybe Text)

Phone number for reporting abuse.

Type: String

gddrAdminContact :: Lens' GetDomainDetailResponse ContactDetail

Provides details about the domain administrative contact.

Type: Complex

Children: FirstName, MiddleName, LastName, ContactType, OrganizationName, AddressLine1, AddressLine2, City, State, CountryCode, ZipCode, PhoneNumber, Email, Fax, ExtraParams

gddrAdminPrivacy :: Lens' GetDomainDetailResponse (Maybe Bool)

Specifies whether contact information for the admin contact is concealed from WHOIS queries. If the value is true, WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter.

Type: Boolean

gddrAutoRenew :: Lens' GetDomainDetailResponse (Maybe Bool)

Specifies whether the domain registration is set to renew automatically.

Type: Boolean

gddrCreationDate :: Lens' GetDomainDetailResponse (Maybe UTCTime)

The date when the domain was created as found in the response to a WHOIS query. The date format is Unix time.

gddrDnsSec :: Lens' GetDomainDetailResponse (Maybe Text)

Reserved for future use.

gddrDomainName :: Lens' GetDomainDetailResponse Text

The name of a domain.

Type: String

gddrExpirationDate :: Lens' GetDomainDetailResponse (Maybe UTCTime)

The date when the registration for the domain is set to expire. The date format is Unix time.

gddrNameservers :: Lens' GetDomainDetailResponse [Nameserver]

The name of the domain.

Type: String

gddrRegistrantContact :: Lens' GetDomainDetailResponse ContactDetail

Provides details about the domain registrant.

Type: Complex

Children: FirstName, MiddleName, LastName, ContactType, OrganizationName, AddressLine1, AddressLine2, City, State, CountryCode, ZipCode, PhoneNumber, Email, Fax, ExtraParams

gddrRegistrantPrivacy :: Lens' GetDomainDetailResponse (Maybe Bool)

Specifies whether contact information for the registrant contact is concealed from WHOIS queries. If the value is true, WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter.

Type: Boolean

gddrRegistrarName :: Lens' GetDomainDetailResponse (Maybe Text)

Name of the registrar of the domain as identified in the registry. Amazon Route 53 domains are registered by registrar Gandi. The value is '"GANDI SAS"'.

Type: String

gddrRegistrarUrl :: Lens' GetDomainDetailResponse (Maybe Text)

Web address of the registrar.

Type: String

gddrRegistryDomainId :: Lens' GetDomainDetailResponse (Maybe Text)

Reserved for future use.

gddrReseller :: Lens' GetDomainDetailResponse (Maybe Text)

Reseller of the domain. Domains registered or transferred using Amazon Route 53 domains will have 'Amazon' as the reseller.

Type: String

gddrStatusList :: Lens' GetDomainDetailResponse [Text]

An array of domain name status codes, also known as Extensible Provisioning Protocol (EPP) status codes.

ICANN, the organization that maintains a central database of domain names, has developed a set of domain name status codes that tell you the status of a variety of operations on a domain name, for example, registering a domain name, transferring a domain name to another registrar, renewing the registration for a domain name, and so on. All registrars use this same set of status codes.

For a current list of domain name status codes and an explanation of what each code means, go to the ICANN website and search for 'epp status codes'. (Search on the ICANN website; web searches sometimes return an old version of the document.)

Type: Array of String

gddrTechContact :: Lens' GetDomainDetailResponse ContactDetail

Provides details about the domain technical contact.

Type: Complex

Children: FirstName, MiddleName, LastName, ContactType, OrganizationName, AddressLine1, AddressLine2, City, State, CountryCode, ZipCode, PhoneNumber, Email, Fax, ExtraParams

gddrTechPrivacy :: Lens' GetDomainDetailResponse (Maybe Bool)

Specifies whether contact information for the tech contact is concealed from WHOIS queries. If the value is true, WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter.

Type: Boolean

gddrUpdatedDate :: Lens' GetDomainDetailResponse (Maybe UTCTime)

The last updated date of the domain as found in the response to a WHOIS query. The date format is Unix time.

gddrWhoIsServer :: Lens' GetDomainDetailResponse (Maybe Text)

The fully qualified name of the WHOIS server that can answer the WHOIS query for the domain.

Type: String