Network.AWS.Route53Domains.GetDomainSuggestions

Description

The GetDomainSuggestions operation returns a list of suggested domain names given a string, which can either be a domain name or simply a word or phrase (without spaces).

Creating a Request

getDomainSuggestions

Arguments

:: Text
gdsDomainName

-> Int
gdsSuggestionCount

-> Bool
gdsOnlyAvailable

-> GetDomainSuggestions

Creates a value of GetDomainSuggestions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gdsDomainName - A domain name that you want to use as the basis for a list of possible domain names. The domain name must contain a top-level domain (TLD), such as .com, that Amazon Route 53 supports. For a list of TLDs, see Domains that You Can Register with Amazon Route 53 in the Amazon Route 53 Developer Guide .

gdsSuggestionCount - The number of suggested domain names that you want Amazon Route 53 to return.

gdsOnlyAvailable - If OnlyAvailable is true , Amazon Route 53 returns only domain names that are available. If OnlyAvailable is false , Amazon Route 53 returns domain names without checking whether they're available to be registered. To determine whether the domain is available, you can call checkDomainAvailability for each suggestion.

data GetDomainSuggestions

See: getDomainSuggestions smart constructor.

Request Lenses

gdsDomainName :: Lens' GetDomainSuggestions Text

A domain name that you want to use as the basis for a list of possible domain names. The domain name must contain a top-level domain (TLD), such as .com, that Amazon Route 53 supports. For a list of TLDs, see Domains that You Can Register with Amazon Route 53 in the Amazon Route 53 Developer Guide .

gdsSuggestionCount :: Lens' GetDomainSuggestions Int

The number of suggested domain names that you want Amazon Route 53 to return.

gdsOnlyAvailable :: Lens' GetDomainSuggestions Bool

If OnlyAvailable is true , Amazon Route 53 returns only domain names that are available. If OnlyAvailable is false , Amazon Route 53 returns domain names without checking whether they're available to be registered. To determine whether the domain is available, you can call checkDomainAvailability for each suggestion.

Destructuring the Response

getDomainSuggestionsResponse

Arguments

:: Int
gdsrsResponseStatus

-> GetDomainSuggestionsResponse

Creates a value of GetDomainSuggestionsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gdsrsSuggestionsList - A list of possible domain names. If you specified true for OnlyAvailable in the request, the list contains only domains that are available for registration.

gdsrsResponseStatus - -- | The response status code.

data GetDomainSuggestionsResponse

See: getDomainSuggestionsResponse smart constructor.

Response Lenses

gdsrsSuggestionsList :: Lens' GetDomainSuggestionsResponse [DomainSuggestion]

A list of possible domain names. If you specified true for OnlyAvailable in the request, the list contains only domains that are available for registration.

gdsrsResponseStatus :: Lens' GetDomainSuggestionsResponse Int

- | The response status code.