Network.AWS.Route53Domains.GetOperationDetail

Description

This operation returns the current status of an operation that is not completed.

http://docs.aws.amazon.com/Route53/latest/APIReference/api-GetOperationDetail.html

Request

data GetOperationDetail

Request constructor

getOperationDetail

Arguments

:: Text
godOperationId

-> GetOperationDetail

GetOperationDetail constructor.

The fields accessible through corresponding lenses are:

godOperationId :: Text

Request lenses

godOperationId :: Lens' GetOperationDetail Text

The identifier for the operation for which you want to get the status. Amazon Route 53 returned the identifier in the response to the original request.

Type: String

Default: None

Required: Yes

Response

data GetOperationDetailResponse

Response constructor

getOperationDetailResponse :: GetOperationDetailResponse

GetOperationDetailResponse constructor.

The fields accessible through corresponding lenses are:

godrDomainName :: Maybe Text

godrMessage :: Maybe Text

godrOperationId :: Maybe Text

godrStatus :: Maybe OperationStatus

godrSubmittedDate :: Maybe UTCTime

godrType :: Maybe OperationType

Response lenses

godrDomainName :: Lens' GetOperationDetailResponse (Maybe Text)

The name of a domain.

Type: String

godrMessage :: Lens' GetOperationDetailResponse (Maybe Text)

Detailed information on the status including possible errors.

Type: String

godrOperationId :: Lens' GetOperationDetailResponse (Maybe Text)

The identifier for the operation.

Type: String

godrStatus :: Lens' GetOperationDetailResponse (Maybe OperationStatus)

The current status of the requested operation in the system.

Type: String

godrSubmittedDate :: Lens' GetOperationDetailResponse (Maybe UTCTime)

The date when the request was submitted.

godrType :: Lens' GetOperationDetailResponse (Maybe OperationType)

The type of operation that was requested.

Type: String