Network.AWS.Route53Domains.ListDomains

Description

This operation returns all the domain names registered with Amazon Route 53 for the current AWS account.

http://docs.aws.amazon.com/Route53/latest/APIReference/api-ListDomains.html

Request

data ListDomains

Request constructor

listDomains :: ListDomains

ListDomains constructor.

The fields accessible through corresponding lenses are:

ldMarker :: Maybe Text

ldMaxItems :: Maybe Int

Request lenses

ldMarker :: Lens' ListDomains (Maybe Text)

For an initial request for a list of domains, omit this element. If the number of domains that are associated with the current AWS account is greater than the value that you specified for MaxItems, you can use Marker to return additional domains. Get the value of NextPageMarker from the previous response, and submit another request that includes the value of NextPageMarker in the Marker element.

Type: String

Default: None

Constraints: The marker must match the value specified in the previous request.

Required: No

ldMaxItems :: Lens' ListDomains (Maybe Int)

Number of domains to be returned.

Type: Integer

Default: 20

Constraints: A numeral between 1 and 100.

Required: No

Response

data ListDomainsResponse

Response constructor

listDomainsResponse :: ListDomainsResponse

ListDomainsResponse constructor.

The fields accessible through corresponding lenses are:

ldrDomains :: [DomainSummary]

ldrNextPageMarker :: Maybe Text

Response lenses

ldrDomains :: Lens' ListDomainsResponse [DomainSummary]

A summary of domains.

Type: Complex type containing a list of domain summaries.

Children: AutoRenew, DomainName, Expiry, TransferLock

ldrNextPageMarker :: Lens' ListDomainsResponse (Maybe Text)

If there are more domains than you specified for MaxItems in the request, submit another request and include the value of NextPageMarker in the value of Marker.

Type: String

Parent: Operations