Network.AWS.Route53Domains.ListOperations

Description

This operation returns the operation IDs of operations that are not yet complete. ListOperations Example POST HTTP1.1 host:route53domains.us-east-1.amazonaws.com x-amz-date:20140711T205230Z authorization:AWS4-HMAC-SHA256 Credential=AKIAIOSFODNN7EXAMPLE20140711us-east-1route53domainsaws4_request,

SignedHeaders=content-length;content-type;host;user-agent;x-amz-date;x-amz-target, Signature=[calculated-signature] x-amz-target:Route53Domains_v20140515.ListOperations user-agent:aws-sdk-java1.8.3 Linux2.6.18-164.el5PAE Java_HotSpot (TM )_Server_VM24.60-b091.7.0_60 content-type:application/x-amz-json-1.1 content-length:[number of characters in the JSON string] { MaxItems : 2 } HTTP/1.1 200 Content-Length:[number of characters in the JSON string] { Operations:[ { OperationId:"4ced3d4a-e011-45ee-b94f-1e2d73477562", Status:WORKFLOW_IN_PROGRESS, SubmittedDate:1403548979.088, Type:CHANGE_PRIVACY_PROTECTION }, { OperationId:"2e3ac45b-89b3-47ea-a042-f56dcd1b6883", Status:WORKFLOW_IN_PROGRESS, SubmittedDate:1403548986.429, Type:DOMAIN_LOCK } ] }.

Request

data ListOperations

The ListOperations request includes the following elements.

Request constructor

listOperations :: ListOperations

Smart constructor for the minimum required parameters to construct a valid ListOperations request.

The fields accessible through corresponding lenses are:

Marker :: Maybe Text

MaxItems :: Maybe Integer

Request lenses

loMarker :: Lens' ListOperations (Maybe Text)

For an initial request for a list of operations, omit this element. If the number of operations that are not yet complete is greater than the value that you specified for MaxItems, you can use Marker to return additional operations. Get the value of NextPageMarker from the previous response, and submit another request that includes the value of NextPageMarker in the Marker element. Type: String Default: None Required: No.

loMaxItems :: Lens' ListOperations (Maybe Integer)

Number of domains to be returned. Type: Integer Default: 20 Constraints: A value between 1 and 100. Required: No.

Response

data ListOperationsResponse

The ListOperations response includes the following elements.

Response constructor

listOperationsResponse

Arguments

:: [OperationSummary]
lorOperations

-> ListOperationsResponse

Smart constructor for the minimum required parameters to construct a valid ListOperationsResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Operations :: [OperationSummary]

NextPageMarker :: Maybe Text

Response lenses

lorOperations :: Lens' ListOperationsResponse [OperationSummary]

Lists summaries of the operations. Type: Complex type containing a list of operation summaries Children: OperationId, Status, SubmittedDate, Type.

lorNextPageMarker :: Lens' ListOperationsResponse (Maybe Text)

If there are more operations than you specified for MaxItems in the request, submit another request and include the value of NextPageMarker in the value of Marker. Type: String Parent: Operations.