Network.AWS.Route53Domains.RegisterDomain

Description

This operation registers a domain. Domains are registered by the AWS registrar partner, Gandi. For some top-level domains (TLDs), this operation requires extra parameters.

When you register a domain, Amazon Route 53 does the following:

Creates a Amazon Route 53 hosted zone that has the same name as the domain. Amazon Route 53 assigns four name servers to your hosted zone and automatically updates your domain registration with the names of these name servers. Enables autorenew, so your domain registration will renew automatically each year. We'll notify you in advance of the renewal date so you can choose whether to renew the registration. Optionally enables privacy protection, so WHOIS queries return contact information for our registrar partner, Gandi, instead of the information you entered for registrant, admin, and tech contacts. If registration is successful, returns an operation ID that you can use to track the progress and completion of the action. If the request is not completed successfully, the domain registrant is notified by email. Charges your AWS account an amount based on the top-level domain. For more information, see Amazon Route 53 Pricing.

http://docs.aws.amazon.com/Route53/latest/APIReference/api-RegisterDomain.html

Request

data RegisterDomain

Request constructor

registerDomain

Arguments

:: Text
rdDomainName

-> Natural
rdDurationInYears

-> ContactDetail
rdAdminContact

-> ContactDetail
rdRegistrantContact

-> ContactDetail
rdTechContact

-> RegisterDomain

RegisterDomain constructor.

The fields accessible through corresponding lenses are:

rdAdminContact :: ContactDetail

rdAutoRenew :: Maybe Bool

rdDomainName :: Text

rdDurationInYears :: Natural

rdIdnLangCode :: Maybe Text

rdPrivacyProtectAdminContact :: Maybe Bool

rdPrivacyProtectRegistrantContact :: Maybe Bool

rdPrivacyProtectTechContact :: Maybe Bool

rdRegistrantContact :: ContactDetail

rdTechContact :: ContactDetail

Request lenses

rdAdminContact :: Lens' RegisterDomain ContactDetail

Provides detailed contact information.

Type: Complex

Children: FirstName, MiddleName, LastName, ContactType, OrganizationName, AddressLine1, AddressLine2, City, State, CountryCode, ZipCode, PhoneNumber, Email, Fax, ExtraParams

Required: Yes

rdAutoRenew :: Lens' RegisterDomain (Maybe Bool)

Indicates whether the domain will be automatically renewed (true) or not (false). Autorenewal only takes effect after the account is charged.

Type: Boolean

Valid values: true | false

Default: true

Required: No

rdDomainName :: Lens' RegisterDomain Text

The name of a domain.

Type: String

Default: None

Constraints: The domain name can contain only the letters a through z, the numbers 0 through 9, and hyphen (-). Internationalized Domain Names are not supported.

Required: Yes

rdDurationInYears :: Lens' RegisterDomain Natural

The number of years the domain will be registered. Domains are registered for a minimum of one year. The maximum period depends on the top-level domain.

Type: Integer

Default: 1

Valid values: Integer from 1 to 10

Required: Yes

rdIdnLangCode :: Lens' RegisterDomain (Maybe Text)

Reserved for future use.

rdPrivacyProtectAdminContact :: Lens' RegisterDomain (Maybe Bool)

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter.

Type: Boolean

Default: true

Valid values: true | false

Required: No

rdPrivacyProtectRegistrantContact :: Lens' RegisterDomain (Maybe Bool)

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter.

Type: Boolean

Default: true

Valid values: true | false

Required: No

rdPrivacyProtectTechContact :: Lens' RegisterDomain (Maybe Bool)

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter.

Type: Boolean

Default: true

Valid values: true | false

Required: No

rdRegistrantContact :: Lens' RegisterDomain ContactDetail

Provides detailed contact information.

Type: Complex

Children: FirstName, MiddleName, LastName, ContactType, OrganizationName, AddressLine1, AddressLine2, City, State, CountryCode, ZipCode, PhoneNumber, Email, Fax, ExtraParams

Required: Yes

rdTechContact :: Lens' RegisterDomain ContactDetail

Provides detailed contact information.

Type: Complex

Children: FirstName, MiddleName, LastName, ContactType, OrganizationName, AddressLine1, AddressLine2, City, State, CountryCode, ZipCode, PhoneNumber, Email, Fax, ExtraParams

Required: Yes

Response

data RegisterDomainResponse

Response constructor

registerDomainResponse

Arguments

:: Text
rdrOperationId

-> RegisterDomainResponse

RegisterDomainResponse constructor.

The fields accessible through corresponding lenses are:

rdrOperationId :: Text

Response lenses

rdrOperationId :: Lens' RegisterDomainResponse Text

Identifier for tracking the progress of the request. To use this ID to query the operation status, use GetOperationDetail.

Type: String

Default: None

Constraints: Maximum 255 characters.