Network.AWS.Route53Domains.ResendContactReachabilityEmail

Description

For operations that require confirmation that the email address for the registrant contact is valid, such as registering a new domain, this operation resends the confirmation email to the current email address for the registrant contact.

Creating a Request

resendContactReachabilityEmail :: ResendContactReachabilityEmail

Creates a value of ResendContactReachabilityEmail with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rcreDomainName - The name of the domain for which you want Amazon Route 53 to resend a confirmation email to the registrant contact.

data ResendContactReachabilityEmail

See: resendContactReachabilityEmail smart constructor.

Request Lenses

rcreDomainName :: Lens' ResendContactReachabilityEmail (Maybe Text)

The name of the domain for which you want Amazon Route 53 to resend a confirmation email to the registrant contact.

Destructuring the Response

resendContactReachabilityEmailResponse

Arguments

:: Int
rcrersResponseStatus

-> ResendContactReachabilityEmailResponse

Creates a value of ResendContactReachabilityEmailResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rcrersDomainName - The domain name for which you requested a confirmation email.

rcrersEmailAddress - The email address for the registrant contact at the time that we sent the verification email.

rcrersIsAlreadyVerified - True if the email address for the registrant contact has already been verified, and false otherwise. If the email address has already been verified, we don't send another confirmation email.

rcrersResponseStatus - -- | The response status code.

data ResendContactReachabilityEmailResponse

See: resendContactReachabilityEmailResponse smart constructor.

Response Lenses

rcrersDomainName :: Lens' ResendContactReachabilityEmailResponse (Maybe Text)

The domain name for which you requested a confirmation email.

rcrersEmailAddress :: Lens' ResendContactReachabilityEmailResponse (Maybe Text)

The email address for the registrant contact at the time that we sent the verification email.

rcrersIsAlreadyVerified :: Lens' ResendContactReachabilityEmailResponse (Maybe Bool)

True if the email address for the registrant contact has already been verified, and false otherwise. If the email address has already been verified, we don't send another confirmation email.

rcrersResponseStatus :: Lens' ResendContactReachabilityEmailResponse Int

- | The response status code.