Network.AWS.Route53Domains.Types

Service

data Route53Domains

Version 2014-05-15 of the Amazon Route 53 Domains service.

Error

data JSONError :: *

DomainSummary

data DomainSummary

domainSummary

Arguments

:: Text
dsDomainName

-> DomainSummary

DomainSummary constructor.

The fields accessible through corresponding lenses are:

dsAutoRenew :: Maybe Bool

dsDomainName :: Text

dsExpiry :: Maybe UTCTime

dsTransferLock :: Maybe Bool

dsAutoRenew :: Lens' DomainSummary (Maybe Bool)

Indicates whether the domain is automatically renewed upon expiration.

Type: Boolean

Valid values: True | False

dsDomainName :: Lens' DomainSummary Text

The name of a domain.

Type: String

dsExpiry :: Lens' DomainSummary (Maybe UTCTime)

Expiration date of the domain in Coordinated Universal Time (UTC).

Type: Long

dsTransferLock :: Lens' DomainSummary (Maybe Bool)

Indicates whether a domain is locked from unauthorized transfer to another party.

Type: Boolean

Valid values: True | False

Tag

data Tag

tag :: Tag

Tag constructor.

The fields accessible through corresponding lenses are:

tagKey :: Maybe Text

tagValue :: Maybe Text

tagKey :: Lens' Tag (Maybe Text)

The key (name) of a tag.

Type: String

Default: None

Valid values: A-Z, a-z, 0-9, space, ".:/=+-%@"

Constraints: Each key can be 1-128 characters long.

Required: Yes

tagValue :: Lens' Tag (Maybe Text)

The value of a tag.

Type: String

Default: None

Valid values: A-Z, a-z, 0-9, space, ".:/=+-%@"

Constraints: Each value can be 0-256 characters long.

Required: Yes

ExtraParamName

data ExtraParamName

Constructors

AuIdNumber
AU_ID_NUMBER

AuIdType
AU_ID_TYPE

BirthCity
BIRTH_CITY

BirthCountry
BIRTH_COUNTRY

BirthDateInYyyyMmDd
BIRTH_DATE_IN_YYYY_MM_DD

BirthDepartment
BIRTH_DEPARTMENT

BrandNumber
BRAND_NUMBER

CaLegalType
CA_LEGAL_TYPE

DocumentNumber
DOCUMENT_NUMBER

DunsNumber
DUNS_NUMBER

EsIdentification
ES_IDENTIFICATION

EsIdentificationType
ES_IDENTIFICATION_TYPE

EsLegalForm
ES_LEGAL_FORM

FiBusinessNumber
FI_BUSINESS_NUMBER

FiIdNumber
FI_ID_NUMBER

ItPin
IT_PIN

RuPassportData
RU_PASSPORT_DATA

SeIdNumber
SE_ID_NUMBER

SgIdNumber
SG_ID_NUMBER

VatNumber
VAT_NUMBER

Nameserver

data Nameserver

nameserver

Arguments

:: Text
nName

-> Nameserver

Nameserver constructor.

The fields accessible through corresponding lenses are:

nGlueIps :: [Text]

nName :: Text

nGlueIps :: Lens' Nameserver [Text]

Glue IP address of a name server entry. Glue IP addresses are required only when the name of the name server is a subdomain of the domain. For example, if your domain is example.com and the name server for the domain is ns.example.com, you need to specify the IP address for ns.example.com.

Type: List of IP addresses.

Constraints: The list can contain only one IPv4 and one IPv6 address.

Parent: Nameservers

nName :: Lens' Nameserver Text

The fully qualified host name of the name server.

Type: String

Constraint: Maximum 255 characterss

Parent: Nameservers

OperationStatus

data OperationStatus

Constructors

Error
ERROR

Failed
FAILED

InProgress
IN_PROGRESS

Submitted
SUBMITTED

Successful
SUCCESSFUL

DomainAvailability

data DomainAvailability

Constructors

Available
AVAILABLE

AvailablePreorder
AVAILABLE_PREORDER

AvailableReserved
AVAILABLE_RESERVED

Reserved
RESERVED

Unavailable
UNAVAILABLE

UnavailablePremium
UNAVAILABLE_PREMIUM

UnavailableRestricted
UNAVAILABLE_RESTRICTED

OperationType

data OperationType

Constructors

OTChangePrivacyProtection
CHANGE_PRIVACY_PROTECTION

OTDeleteDomain
DELETE_DOMAIN

OTDomainLock
DOMAIN_LOCK

OTRegisterDomain
REGISTER_DOMAIN

OTTransferInDomain
TRANSFER_IN_DOMAIN

OTUpdateDomainContact
UPDATE_DOMAIN_CONTACT

OTUpdateNameserver
UPDATE_NAMESERVER

CountryCode

data CountryCode

Constructors

Ad
AD

Ae
AE

Af
AF

Ag
AG

Ai
AI

Al
AL

Am
AM

An
AN

Ao
AO

Aq
AQ

Ar
AR

As
AS

At
AT

Au
AU

Aw
AW

Az
AZ

Ba
BA

Bb
BB

Bd
BD

Be
BE

Bf
BF

Bg
BG

Bh
BH

Bi
BI

Bj
BJ

Bl
BL

Bm
BM

Bn
BN

Bo
BO

Br
BR

Bs
BS

Bt
BT

Bw
BW

By
BY

Bz
BZ

Ca
CA

Cc
CC

Cd
CD

Cf
CF

Cg
CG

Ch
CH

Ci
CI

Ck
CK

Cl
CL

Cm
CM

Cn
CN

Co
CO

Cr
CR

Cu
CU

Cv
CV

Cx
CX

Cy
CY

Cz
CZ

De
DE

Dj
DJ

Dk
DK

Dm
DM

Do'
DO

Dz
DZ

Ec
EC

Ee
EE

Eg
EG

Er
ER

Es
ES

Et
ET

Fi
FI

Fj
FJ

Fk
FK

Fm
FM

Fo
FO

Fr
FR

Ga
GA

Gb
GB

Gd
GD

Ge
GE

Gh
GH

Gi
GI

Gl
GL

Gm
GM

Gn
GN

Gq
GQ

Gr
GR

Gt
GT

Gu
GU

Gw
GW

Gy
GY

Hk
HK

Hn
HN

Hr
HR

Ht
HT

Hu
HU

Id
ID

Ie
IE

Il
IL

Im
IM

In'
IN

Iq
IQ

Ir
IR

Is
IS

It
IT

Jm
JM

Jo
JO

Jp
JP

Ke
KE

Kg
KG

Kh
KH

Ki
KI

Km
KM

Kn
KN

Kp
KP

Kr
KR

Kw
KW

Ky
KY

Kz
KZ

La
LA

Lb
LB

Lc
LC

Li
LI

Lk
LK

Lr
LR

Ls
LS

Lt
LT

Lu
LU

Lv
LV

Ly
LY

Ma
MA

Mc
MC

Md
MD

Me
ME

Mf
MF

Mg
MG

Mh
MH

Mk
MK

Ml
ML

Mm
MM

Mn
MN

Mo
MO

Mp
MP

Mr
MR

Ms
MS

Mt
MT

Mu
MU

Mv
MV

Mw
MW

Mx
MX

My
MY

Mz
MZ

Na
NA

Nc
NC

Ne
NE

Ng
NG

Ni
NI

Nl
NL

No
NO

Np
NP

Nr
NR

Nu
NU

Nz
NZ

Om
OM

Pa
PA

Pe
PE

Pf
PF

Pg
PG

Ph
PH

Pk
PK

Pl
PL

Pm
PM

Pn
PN

Pr
PR

Pt
PT

Pw
PW

Py
PY

Qa
QA

Ro
RO

Rs
RS

Ru
RU

Rw
RW

Sa
SA

Sb
SB

Sc
SC

Sd
SD

Se
SE

Sg
SG

Sh
SH

Si
SI

Sk
SK

Sl
SL

Sm
SM

Sn
SN

So
SO

Sr
SR

St
ST

Sv
SV

Sy
SY

Sz
SZ

Tc
TC

Td
TD

Tg
TG

Th
TH

Tj
TJ

Tk
TK

Tl
TL

Tm
TM

Tn
TN

To
TO

Tr
TR

Tt
TT

Tv
TV

Tw
TW

Tz
TZ

Ua
UA

Ug
UG

Us
US

Uy
UY

Uz
UZ

Va
VA

Vc
VC

Ve
VE

Vg
VG

Vi
VI

Vn
VN

Vu
VU

Wf
WF

Ws
WS

Ye
YE

Yt
YT

Za
ZA

Zm
ZM

Zw
ZW

ExtraParam

data ExtraParam

extraParam

Arguments

:: ExtraParamName
epName

-> Text
epValue

-> ExtraParam

ExtraParam constructor.

The fields accessible through corresponding lenses are:

epName :: ExtraParamName

epValue :: Text

epName :: Lens' ExtraParam ExtraParamName

Name of the additional parameter required by the top-level domain.

Type: String

Default: None

Valid values: DUNS_NUMBER | BRAND_NUMBER | BIRTH_DEPARTMENT | BIRTH_DATE_IN_YYYY_MM_DD | BIRTH_COUNTRY | BIRTH_CITY | DOCUMENT_NUMBER | AU_ID_NUMBER | AU_ID_TYPE | CA_LEGAL_TYPE | ES_IDENTIFICATION | ES_IDENTIFICATION_TYPE | ES_LEGAL_FORM | FI_BUSINESS_NUMBER | FI_ID_NUMBER | IT_PIN | RU_PASSPORT_DATA | SE_ID_NUMBER | SG_ID_NUMBER | VAT_NUMBER

Parent: ExtraParams

Required: Yes

epValue :: Lens' ExtraParam Text

Values corresponding to the additional parameter names required by some top-level domains.

Type: String

Default: None

Constraints: Maximum 2048 characters.

Parent: ExtraParams

Required: Yes

ContactType

data ContactType

Constructors

CTAssociation
ASSOCIATION

CTCompany
COMPANY

CTPerson
PERSON

CTPublicBody
PUBLIC_BODY

CTReseller
RESELLER

ContactDetail

data ContactDetail

contactDetail :: ContactDetail

ContactDetail constructor.

The fields accessible through corresponding lenses are:

cdAddressLine1 :: Maybe Text

cdAddressLine2 :: Maybe Text

cdCity :: Maybe Text

cdContactType :: Maybe ContactType

cdCountryCode :: Maybe CountryCode

cdEmail :: Maybe Text

cdExtraParams :: [ExtraParam]

cdFax :: Maybe Text

cdFirstName :: Maybe Text

cdLastName :: Maybe Text

cdOrganizationName :: Maybe Text

cdPhoneNumber :: Maybe Text

cdState :: Maybe Text

cdZipCode :: Maybe Text

cdAddressLine1 :: Lens' ContactDetail (Maybe Text)

First line of the contact's address.

Type: String

Default: None

Constraints: Maximum 255 characters.

Parents: RegistrantContact, AdminContact, TechContact

Required: Yes

cdAddressLine2 :: Lens' ContactDetail (Maybe Text)

Second line of contact's address, if any.

Type: String

Default: None

Constraints: Maximum 255 characters.

Parents: RegistrantContact, AdminContact, TechContact

Required: No

cdCity :: Lens' ContactDetail (Maybe Text)

The city of the contact's address.

Type: String

Default: None

Constraints: Maximum 255 characters.

Parents: RegistrantContact, AdminContact, TechContact

Required: Yes

cdContactType :: Lens' ContactDetail (Maybe ContactType)

Indicates whether the contact is a person, company, association, or public organization. If you choose an option other than PERSON, you must enter an organization name, and you can't enable privacy protection for the contact.

Type: String

Default: None

Constraints: Maximum 255 characters.

Valid values: PERSON | COMPANY | ASSOCIATION | PUBLIC_BODY

Parents: RegistrantContact, AdminContact, TechContact

Required: Yes

cdCountryCode :: Lens' ContactDetail (Maybe CountryCode)

Code for the country of the contact's address.

Type: String

Default: None

Constraints: Maximum 255 characters.

Parents: RegistrantContact, AdminContact, TechContact

Required: Yes

cdEmail :: Lens' ContactDetail (Maybe Text)

Email address of the contact.

Type: String

Default: None

Constraints: Maximum 254 characters.

Parents: RegistrantContact, AdminContact, TechContact

Required: Yes

cdExtraParams :: Lens' ContactDetail [ExtraParam]

A list of name-value pairs for parameters required by certain top-level domains.

Type: Complex

Default: None

Parents: RegistrantContact, AdminContact, TechContact

Children: Name, Value

Required: No

cdFax :: Lens' ContactDetail (Maybe Text)

Fax number of the contact.

Type: String

Default: None

Constraints: Phone number must be specified in the format "+[country dialing code].[number including any area code]". For example, a US phone number might appear as '"+1.1234567890"'.

Parents: RegistrantContact, AdminContact, TechContact

Required: No

cdFirstName :: Lens' ContactDetail (Maybe Text)

First name of contact.

Type: String

Default: None

Constraints: Maximum 255 characters.

Parents: RegistrantContact, AdminContact, TechContact

Required: Yes

cdLastName :: Lens' ContactDetail (Maybe Text)

Last name of contact.

Type: String

Default: None

Constraints: Maximum 255 characters.

Parents: RegistrantContact, AdminContact, TechContact

Required: Yes

cdOrganizationName :: Lens' ContactDetail (Maybe Text)

Name of the organization for contact types other than PERSON.

Type: String

Default: None

Constraints: Maximum 255 characters. Contact type must not be PERSON.

Parents: RegistrantContact, AdminContact, TechContact

Required: No

cdPhoneNumber :: Lens' ContactDetail (Maybe Text)

The phone number of the contact.

Type: String

Default: None

Constraints: Phone number must be specified in the format "+[country dialing code].[number including any area code>]". For example, a US phone number might appear as '"+1.1234567890"'.

Parents: RegistrantContact, AdminContact, TechContact

Required: Yes

cdState :: Lens' ContactDetail (Maybe Text)

The state or province of the contact's city.

Type: String

Default: None

Constraints: Maximum 255 characters.

Parents: RegistrantContact, AdminContact, TechContact

Required: No

cdZipCode :: Lens' ContactDetail (Maybe Text)

The zip or postal code of the contact's address.

Type: String

Default: None

Constraints: Maximum 255 characters.

Parents: RegistrantContact, AdminContact, TechContact

Required: No

OperationSummary

data OperationSummary

operationSummary

Arguments

:: Text
osOperationId

-> OperationStatus
osStatus

-> OperationType
osType

-> UTCTime
osSubmittedDate

-> OperationSummary

OperationSummary constructor.

The fields accessible through corresponding lenses are:

osOperationId :: Text

osStatus :: OperationStatus

osSubmittedDate :: UTCTime

osType :: OperationType

osOperationId :: Lens' OperationSummary Text

Identifier returned to track the requested action.

Type: String

osStatus :: Lens' OperationSummary OperationStatus

The current status of the requested operation in the system.

Type: String

osSubmittedDate :: Lens' OperationSummary UTCTime

The date when the request was submitted.

osType :: Lens' OperationSummary OperationType

Type of the action requested.

Type: String

Valid values: REGISTER_DOMAIN | DELETE_DOMAIN | TRANSFER_IN_DOMAIN | UPDATE_DOMAIN_CONTACT | UPDATE_NAMESERVER | CHANGE_PRIVACY_PROTECTION | DOMAIN_LOCK