Network.AWS.Route53Domains.UpdateDomainContact

Description

This operation updates the contact information for a particular domain. Information for at least one contact (registrant, administrator, or technical) must be supplied for update.

If the update is successful, this method returns an operation ID that you can use to track the progress and completion of the action. If the request is not completed successfully, the domain registrant will be notified by email.

Creating a Request

updateDomainContact

Arguments

:: Text
udcDomainName

-> UpdateDomainContact

Creates a value of UpdateDomainContact with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

udcRegistrantContact - Provides detailed contact information.

udcAdminContact - Provides detailed contact information.

udcTechContact - Provides detailed contact information.

udcDomainName - The name of the domain that you want to update contact information for.

data UpdateDomainContact

The UpdateDomainContact request includes the following elements.

See: updateDomainContact smart constructor.

Request Lenses

udcRegistrantContact :: Lens' UpdateDomainContact (Maybe ContactDetail)

Provides detailed contact information.

udcAdminContact :: Lens' UpdateDomainContact (Maybe ContactDetail)

Provides detailed contact information.

udcTechContact :: Lens' UpdateDomainContact (Maybe ContactDetail)

Provides detailed contact information.

udcDomainName :: Lens' UpdateDomainContact Text

The name of the domain that you want to update contact information for.

Destructuring the Response

updateDomainContactResponse

Arguments

:: Int
udcrsResponseStatus

-> Text
udcrsOperationId

-> UpdateDomainContactResponse

Creates a value of UpdateDomainContactResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

udcrsResponseStatus - -- | The response status code.

udcrsOperationId - Identifier for tracking the progress of the request. To use this ID to query the operation status, use GetOperationDetail .

data UpdateDomainContactResponse

The UpdateDomainContact response includes the following element.

See: updateDomainContactResponse smart constructor.

Response Lenses

udcrsResponseStatus :: Lens' UpdateDomainContactResponse Int

- | The response status code.

udcrsOperationId :: Lens' UpdateDomainContactResponse Text

Identifier for tracking the progress of the request. To use this ID to query the operation status, use GetOperationDetail .