Network.AWS.Route53Domains.UpdateDomainContactPrivacy

Description

This operation updates the specified domain contact's privacy setting. When the privacy option is enabled, personal information such as postal or email address is hidden from the results of a public WHOIS query. The privacy services are provided by the AWS registrar, Gandi. For more information, see the Gandi privacy features. This operation only affects the privacy of the specified contact type (registrant, administrator, or tech). Successful acceptance returns an operation ID that you can use with GetOperationDetail to track the progress and completion of the action. If the request is not completed successfully, the domain registrant will be notified by email.

http://docs.aws.amazon.com/Route53/latest/APIReference/api-UpdateDomainContactPrivacy.html

Request

data UpdateDomainContactPrivacy

Request constructor

updateDomainContactPrivacy

Arguments

:: Text
udcpDomainName

-> UpdateDomainContactPrivacy

UpdateDomainContactPrivacy constructor.

The fields accessible through corresponding lenses are:

udcpAdminPrivacy :: Maybe Bool

udcpDomainName :: Text

udcpRegistrantPrivacy :: Maybe Bool

udcpTechPrivacy :: Maybe Bool

Request lenses

udcpAdminPrivacy :: Lens' UpdateDomainContactPrivacy (Maybe Bool)

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter. Type: Boolean Default: None Valid values: true | false Required: No.

udcpDomainName :: Lens' UpdateDomainContactPrivacy Text

The name of a domain. Type: String Default: None Constraints: The domain name can contain only the letters a through z, the numbers 0 through 9, and hyphen (-). Internationalized Domain Names are not supported. Required: Yes.

udcpRegistrantPrivacy :: Lens' UpdateDomainContactPrivacy (Maybe Bool)

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter. Type: Boolean Default: None Valid values: true | false Required: No.

udcpTechPrivacy :: Lens' UpdateDomainContactPrivacy (Maybe Bool)

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter. Type: Boolean Default: None Valid values: true | false Required: No.

Response

data UpdateDomainContactPrivacyResponse

Response constructor

updateDomainContactPrivacyResponse

Arguments

:: Text
udcprOperationId

-> UpdateDomainContactPrivacyResponse

UpdateDomainContactPrivacyResponse constructor.

The fields accessible through corresponding lenses are:

udcprOperationId :: Text

Response lenses

udcprOperationId :: Lens' UpdateDomainContactPrivacyResponse Text

Identifier for tracking the progress of the request. To use this ID to query the operation status, use GetOperationDetail. Type: String Default: None Constraints: Maximum 255 characters.