Network.AWS.Route53Domains.V2014_05_15

Description

Pending.

Documentation

checkDomainAvailability

Arguments

:: Text
_cdarDomainName

-> CheckDomainAvailability

Minimum specification for a CheckDomainAvailability request.

data CheckDomainAvailability

Constructors

CheckDomainAvailability

Fields

_cdarDomainName :: Text

The name of a domain. Type: String Default: None Constraints: The domain name can contain only the letters a through z, the numbers 0 through 9, and hyphen (-). Internationalized Domain Names are not supported. Required: Yes.

_cdarIdnLangCode :: Maybe Text

Reserved for future use.

cdarIdnLangCode :: Lens' CheckDomainAvailability (Maybe Text)

cdarDomainName :: Lens' CheckDomainAvailability Text

data CheckDomainAvailabilityResponse

Constructors

CheckDomainAvailabilityResponse

Fields

_cdasAvailability :: DomainAvailability

Whether the domain name is available for registering. You can only register domains designated as AVAILABLE. Type: String Valid values: AVAILABLE – The domain name is available. AVAILABLE_RESERVED – The domain name is reserved under specific conditions. AVAILABLE_PREORDER – The domain name is available and can be preordered. UNAVAILABLE – The domain name is not available. UNAVAILABLE_PREMIUM – The domain name is not available. UNAVAILABLE_RESTRICTED – The domain name is forbidden. RESERVED – The domain name has been reserved for another person or organization.

cdasAvailability :: Iso' CheckDomainAvailabilityResponse DomainAvailability

data DisableDomainTransferLock

Constructors

DisableDomainTransferLock

Fields

_ddtlrDomainName :: Text

The name of a domain. Type: String Default: None Constraints: The domain name can contain only the letters a through z, the numbers 0 through 9, and hyphen (-). Internationalized Domain Names are not supported. Required: Yes.

ddtlrDomainName :: Iso' DisableDomainTransferLock Text

data DisableDomainTransferLockResponse

Constructors

DisableDomainTransferLockResponse

Fields

_ddtlsOperationId :: Text

Identifier for tracking the progress of the request. To use this ID to query the operation status, use GetOperationDetail. Type: String Default: None Constraints: Maximum 255 characters.

ddtlsOperationId :: Iso' DisableDomainTransferLockResponse Text

data EnableDomainTransferLock

Constructors

EnableDomainTransferLock

Fields

_edtlrDomainName :: Text

The name of a domain. Type: String Default: None Constraints: The domain name can contain only the letters a through z, the numbers 0 through 9, and hyphen (-). Internationalized Domain Names are not supported. Required: Yes.

edtlrDomainName :: Iso' EnableDomainTransferLock Text

data EnableDomainTransferLockResponse

Constructors

EnableDomainTransferLockResponse

Fields

_edtlsOperationId :: Text

Identifier for tracking the progress of the request. To use this ID to query the operation status, use GetOperationDetail. Type: String Default: None Constraints: Maximum 255 characters.

edtlsOperationId :: Iso' EnableDomainTransferLockResponse Text

data GetDomainDetail

Constructors

GetDomainDetail

Fields

_gddrDomainName :: Text

The name of a domain. Type: String Default: None Constraints: The domain name can contain only the letters a through z, the numbers 0 through 9, and hyphen (-). Internationalized Domain Names are not supported. Required: Yes.

gddrDomainName :: Iso' GetDomainDetail Text

data GetDomainDetailResponse

Constructors

GetDomainDetailResponse

Fields

_gddsRegistrantContact :: ContactDetail

Provides details about the domain registrant. Type: Complex Children: FirstName, MiddleName, LastName, ContactType, OrganizationName, AddressLine1, AddressLine2, City, State, CountryCode, ZipCode, PhoneNumber, Email, Fax, ExtraParams.

_gddsAdminContact :: ContactDetail

Provides details about the domain administrative contact. Type: Complex Children: FirstName, MiddleName, LastName, ContactType, OrganizationName, AddressLine1, AddressLine2, City, State, CountryCode, ZipCode, PhoneNumber, Email, Fax, ExtraParams.

_gddsTechContact :: ContactDetail

Provides details about the domain technical contact. Type: Complex Children: FirstName, MiddleName, LastName, ContactType, OrganizationName, AddressLine1, AddressLine2, City, State, CountryCode, ZipCode, PhoneNumber, Email, Fax, ExtraParams.

_gddsDomainName :: Text

The name of a domain. Type: String.

_gddsNameservers :: [Nameserver]

The name of the domain. Type: String.

_gddsTechPrivacy :: Maybe Bool

Specifies whether contact information for the tech contact is concealed from WHOIS queries. If the value is true, WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter. Type: Boolean.

_gddsRegistrantPrivacy :: Maybe Bool

Specifies whether contact information for the registrant contact is concealed from WHOIS queries. If the value is true, WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter. Type: Boolean.

_gddsAdminPrivacy :: Maybe Bool

Specifies whether contact information for the admin contact is concealed from WHOIS queries. If the value is true, WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter. Type: Boolean.

_gddsAutoRenew :: Maybe Bool

Specifies whether the domain registration is set to renew automatically. Type: Boolean.

_gddsAbuseContactPhone :: Maybe Text

Phone number for reporting abuse. Type: String.

_gddsDnsSec :: Maybe Text

Reserved for future use.

_gddsStatusList :: [Text]

An array of domain name status codes, also known as Extensible Provisioning Protocol (EPP) status codes. ICANN, the organization that maintains a central database of domain names, has developed a set of domain name status codes that tell you the status of a variety of operations on a domain name, for example, registering a domain name, transferring a domain name to another registrar, renewing the registration for a domain name, and so on. All registrars use this same set of status codes. For a current list of domain name status codes and an explanation of what each code means, go to the ICANN website and search for epp status codes. (Search on the ICANN website; web searches sometimes return an old version of the document.) Type: Array of String.

_gddsAbuseContactEmail :: Maybe Text

Email address to contact to report incorrect contact information for a domain, to report that the domain is being used to send spam, to report that someone is cybersquatting on a domain name, or report some other type of abuse. Type: String.

_gddsRegistrarName :: Maybe Text

Name of the registrar of the domain as identified in the registry. Amazon Route 53 domains are registered by registrar Gandi. The value is "GANDI SAS". Type: String.

_gddsRegistrarUrl :: Maybe Text

Web address of the registrar. Type: String.

_gddsWhoIsServer :: Maybe Text

The fully qualified name of the WHOIS server that can answer the WHOIS query for the domain. Type: String.

_gddsRegistryDomainId :: Maybe Text

Reserved for future use.

_gddsReseller :: Maybe Text

Reseller of the domain. Domains registered or transferred using Amazon Route 53 domains will have Amazon as the reseller. Type: String.

_gddsUpdatedDate :: Maybe ISO8601

The last updated date of the domain as found in the response to a WHOIS query. The date format is Unix time.

_gddsExpirationDate :: Maybe ISO8601

The date when the registration for the domain is set to expire. The date format is Unix time.

_gddsCreationDate :: Maybe ISO8601

The date when the domain was created as found in the response to a WHOIS query. The date format is Unix time.

gddsWhoIsServer :: Lens' GetDomainDetailResponse (Maybe Text)

gddsUpdatedDate :: Lens' GetDomainDetailResponse (Maybe ISO8601)

gddsTechPrivacy :: Lens' GetDomainDetailResponse (Maybe Bool)

gddsTechContact :: Lens' GetDomainDetailResponse ContactDetail

gddsStatusList :: Lens' GetDomainDetailResponse [Text]

gddsReseller :: Lens' GetDomainDetailResponse (Maybe Text)

gddsRegistryDomainId :: Lens' GetDomainDetailResponse (Maybe Text)

gddsRegistrarUrl :: Lens' GetDomainDetailResponse (Maybe Text)

gddsRegistrarName :: Lens' GetDomainDetailResponse (Maybe Text)

gddsRegistrantPrivacy :: Lens' GetDomainDetailResponse (Maybe Bool)

gddsRegistrantContact :: Lens' GetDomainDetailResponse ContactDetail

gddsNameservers :: Lens' GetDomainDetailResponse [Nameserver]

gddsExpirationDate :: Lens' GetDomainDetailResponse (Maybe ISO8601)

gddsDomainName :: Lens' GetDomainDetailResponse Text

gddsDnsSec :: Lens' GetDomainDetailResponse (Maybe Text)

gddsCreationDate :: Lens' GetDomainDetailResponse (Maybe ISO8601)

gddsAutoRenew :: Lens' GetDomainDetailResponse (Maybe Bool)

gddsAdminPrivacy :: Lens' GetDomainDetailResponse (Maybe Bool)

gddsAdminContact :: Lens' GetDomainDetailResponse ContactDetail

gddsAbuseContactPhone :: Lens' GetDomainDetailResponse (Maybe Text)

gddsAbuseContactEmail :: Lens' GetDomainDetailResponse (Maybe Text)

data GetOperationDetail

Constructors

GetOperationDetail

Fields

_godrOperationId :: Text

The identifier for the operation for which you want to get the status. Amazon Route 53 returned the identifier in the response to the original request. Type: String Default: None Required: Yes.

godrOperationId :: Iso' GetOperationDetail Text

data GetOperationDetailResponse

Constructors

GetOperationDetailResponse

Fields

_godsDomainName :: Maybe Text

The name of a domain. Type: String.

_godsMessage :: Maybe Text

Detailed information on the status including possible errors. Type: String.

_godsOperationId :: Maybe Text

The identifier for the operation. Type: String.

_godsStatus :: Maybe OperationStatus

The current status of the requested operation in the system. Type: String.

_godsType :: Maybe OperationType

The type of operation that was requested. Type: String.

_godsSubmittedDate :: Maybe ISO8601

The date when the request was submitted.

godsType :: Lens' GetOperationDetailResponse (Maybe OperationType)

godsSubmittedDate :: Lens' GetOperationDetailResponse (Maybe ISO8601)

godsStatus :: Lens' GetOperationDetailResponse (Maybe OperationStatus)

godsOperationId :: Lens' GetOperationDetailResponse (Maybe Text)

godsMessage :: Lens' GetOperationDetailResponse (Maybe Text)

godsDomainName :: Lens' GetOperationDetailResponse (Maybe Text)

listDomains :: ListDomains

Minimum specification for a ListDomains request.

data ListDomains

Constructors

ListDomains

Fields

_ldrMarker :: Maybe Text

For an initial request for a list of domains, omit this element. If the number of domains that are associated with the current AWS account is greater than the value that you specified for MaxItems, you can use Marker to return additional domains. Get the value of NextPageMarker from the previous response, and submit another request that includes the value of NextPageMarker in the Marker element. Type: String Default: None Constraints: The marker must match the value specified in the previous request. Required: No.

_ldrMaxItems :: Maybe Integer

Number of domains to be returned. Type: Integer Default: 20 Constraints: A numeral between 1 and 100. Required: No.

ldrMaxItems :: Lens' ListDomains (Maybe Integer)

ldrMarker :: Lens' ListDomains (Maybe Text)

data ListDomainsResponse

Constructors

ListDomainsResponse

Fields

_ldsDomains :: [DomainSummary]

A summary of domains. Type: Complex type containing a list of domain summaries. Children: AutoRenew, DomainName, Expiry, TransferLock.

_ldsNextPageMarker :: Maybe Text

If there are more domains than you specified for MaxItems in the request, submit another request and include the value of NextPageMarker in the value of Marker. Type: String Parent: Operations.

ldsNextPageMarker :: Lens' ListDomainsResponse (Maybe Text)

ldsDomains :: Lens' ListDomainsResponse [DomainSummary]

listOperations :: ListOperations

Minimum specification for a ListOperations request.

data ListOperations

Constructors

ListOperations

Fields

_lorMarker :: Maybe Text

For an initial request for a list of operations, omit this element. If the number of operations that are not yet complete is greater than the value that you specified for MaxItems, you can use Marker to return additional operations. Get the value of NextPageMarker from the previous response, and submit another request that includes the value of NextPageMarker in the Marker element. Type: String Default: None Required: No.

_lorMaxItems :: Maybe Integer

Number of domains to be returned. Type: Integer Default: 20 Constraints: A value between 1 and 100. Required: No.

lorMaxItems :: Lens' ListOperations (Maybe Integer)

lorMarker :: Lens' ListOperations (Maybe Text)

data ListOperationsResponse

Constructors

ListOperationsResponse

Fields

_losOperations :: [OperationSummary]

Lists summaries of the operations. Type: Complex type containing a list of operation summaries Children: OperationId, Status, SubmittedDate, Type.

_losNextPageMarker :: Maybe Text

If there are more operations than you specified for MaxItems in the request, submit another request and include the value of NextPageMarker in the value of Marker. Type: String Parent: Operations.

losOperations :: Lens' ListOperationsResponse [OperationSummary]

losNextPageMarker :: Lens' ListOperationsResponse (Maybe Text)

registerDomain

Arguments

:: ContactDetail
_rdrRegistrantContact

-> ContactDetail
_rdrAdminContact

-> ContactDetail
_rdrTechContact

-> Text
_rdrDomainName

-> Integer
_rdrDurationInYears

-> RegisterDomain

Minimum specification for a RegisterDomain request.

data RegisterDomain

Constructors

RegisterDomain

Fields

_rdrRegistrantContact :: ContactDetail

Provides detailed contact information. Type: Complex Children: FirstName, MiddleName, LastName, ContactType, OrganizationName, AddressLine1, AddressLine2, City, State, CountryCode, ZipCode, PhoneNumber, Email, Fax, ExtraParams Required: Yes.

_rdrAdminContact :: ContactDetail

Provides detailed contact information. Type: Complex Children: FirstName, MiddleName, LastName, ContactType, OrganizationName, AddressLine1, AddressLine2, City, State, CountryCode, ZipCode, PhoneNumber, Email, Fax, ExtraParams Required: Yes.

_rdrTechContact :: ContactDetail

Provides detailed contact information. Type: Complex Children: FirstName, MiddleName, LastName, ContactType, OrganizationName, AddressLine1, AddressLine2, City, State, CountryCode, ZipCode, PhoneNumber, Email, Fax, ExtraParams Required: Yes.

_rdrDomainName :: Text

The name of a domain. Type: String Default: None Constraints: The domain name can contain only the letters a through z, the numbers 0 through 9, and hyphen (-). Internationalized Domain Names are not supported. Required: Yes.

_rdrDurationInYears :: Integer

The number of years the domain will be registered. Domains are registered for a minimum of one year. The maximum period depends on the top-level domain. Type: Integer Default: 1 Valid values: Integer from 1 to 10 Required: Yes.

_rdrPrivacyProtectTechContact :: Maybe Bool

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter. Type: Boolean Default: true Valid values: true | false Required: No.

_rdrPrivacyProtectRegistrantContact :: Maybe Bool

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter. Type: Boolean Default: true Valid values: true | false Required: No.

_rdrAutoRenew :: Maybe Bool

Indicates whether the domain will be automatically renewed (true) or not (false). Autorenewal only takes effect after the account is charged. Type: Boolean Valid values: true | false Default: true Required: No.

_rdrPrivacyProtectAdminContact :: Maybe Bool

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter. Type: Boolean Default: true Valid values: true | false Required: No.

_rdrIdnLangCode :: Maybe Text

Reserved for future use.

rdrTechContact :: Lens' RegisterDomain ContactDetail

rdrRegistrantContact :: Lens' RegisterDomain ContactDetail

rdrPrivacyProtectTechContact :: Lens' RegisterDomain (Maybe Bool)

rdrPrivacyProtectRegistrantContact :: Lens' RegisterDomain (Maybe Bool)

rdrPrivacyProtectAdminContact :: Lens' RegisterDomain (Maybe Bool)

rdrIdnLangCode :: Lens' RegisterDomain (Maybe Text)

rdrDurationInYears :: Lens' RegisterDomain Integer

rdrDomainName :: Lens' RegisterDomain Text

rdrAutoRenew :: Lens' RegisterDomain (Maybe Bool)

rdrAdminContact :: Lens' RegisterDomain ContactDetail

data RegisterDomainResponse

Constructors

RegisterDomainResponse

Fields

_rdsOperationId :: Text

Identifier for tracking the progress of the request. To use this ID to query the operation status, use GetOperationDetail. Type: String Default: None Constraints: Maximum 255 characters.

rdsOperationId :: Iso' RegisterDomainResponse Text

data RetrieveDomainAuthCode

Constructors

RetrieveDomainAuthCode

Fields

_rdacrDomainName :: Text

The name of a domain. Type: String Default: None Constraints: The domain name can contain only the letters a through z, the numbers 0 through 9, and hyphen (-). Internationalized Domain Names are not supported. Required: Yes.

rdacrDomainName :: Iso' RetrieveDomainAuthCode Text

data RetrieveDomainAuthCodeResponse

Constructors

RetrieveDomainAuthCodeResponse

Fields

_rdacsAuthCode :: Text

The authorization code for the domain. Type: String.

rdacsAuthCode :: Iso' RetrieveDomainAuthCodeResponse Text

transferDomain

Arguments

:: ContactDetail
_tdrRegistrantContact

-> ContactDetail
_tdrAdminContact

-> ContactDetail
_tdrTechContact

-> Text
_tdrDomainName

-> Integer
_tdrDurationInYears

-> [Nameserver]
_tdrNameservers

-> TransferDomain

Minimum specification for a TransferDomain request.

data TransferDomain

Constructors

TransferDomain

Fields

_tdrRegistrantContact :: ContactDetail

Provides detailed contact information. Type: Complex Children: FirstName, MiddleName, LastName, ContactType, OrganizationName, AddressLine1, AddressLine2, City, State, CountryCode, ZipCode, PhoneNumber, Email, Fax, ExtraParams Required: Yes.

_tdrAdminContact :: ContactDetail

Provides detailed contact information. Type: Complex Children: FirstName, MiddleName, LastName, ContactType, OrganizationName, AddressLine1, AddressLine2, City, State, CountryCode, ZipCode, PhoneNumber, Email, Fax, ExtraParams Required: Yes.

_tdrTechContact :: ContactDetail

Provides detailed contact information. Type: Complex Children: FirstName, MiddleName, LastName, ContactType, OrganizationName, AddressLine1, AddressLine2, City, State, CountryCode, ZipCode, PhoneNumber, Email, Fax, ExtraParams Required: Yes.

_tdrDomainName :: Text

The name of a domain. Type: String Default: None Constraints: The domain name can contain only the letters a through z, the numbers 0 through 9, and hyphen (-). Internationalized Domain Names are not supported. Required: Yes.

_tdrDurationInYears :: Integer

The number of years the domain will be registered. Domains are registered for a minimum of one year. The maximum period depends on the top-level domain. Type: Integer Default: 1 Valid values: Integer from 1 to 10 Required: Yes.

_tdrNameservers :: [Nameserver]

Contains details for the host and glue IP addresses. Type: Complex Children: GlueIps, Name.

_tdrPrivacyProtectTechContact :: Maybe Bool

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter. Type: Boolean Default: true Valid values: true | false Required: No.

_tdrPrivacyProtectRegistrantContact :: Maybe Bool

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter. Type: Boolean Default: true Valid values: true | false Required: No.

_tdrAutoRenew :: Maybe Bool

Indicates whether the domain will be automatically renewed (true) or not (false). Autorenewal only takes effect after the account is charged. Type: Boolean Valid values: true | false Default: true Required: No.

_tdrPrivacyProtectAdminContact :: Maybe Bool

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter. Type: Boolean Default: true Valid values: true | false Required: No.

_tdrAuthCode :: Maybe Text

The authorization code for the domain. You get this value from the current registrar. Type: String Required: Yes.

_tdrIdnLangCode :: Maybe Text

Reserved for future use.

tdrTechContact :: Lens' TransferDomain ContactDetail

tdrRegistrantContact :: Lens' TransferDomain ContactDetail

tdrPrivacyProtectTechContact :: Lens' TransferDomain (Maybe Bool)

tdrPrivacyProtectRegistrantContact :: Lens' TransferDomain (Maybe Bool)

tdrPrivacyProtectAdminContact :: Lens' TransferDomain (Maybe Bool)

tdrNameservers :: Lens' TransferDomain [Nameserver]

tdrIdnLangCode :: Lens' TransferDomain (Maybe Text)

tdrDurationInYears :: Lens' TransferDomain Integer

tdrDomainName :: Lens' TransferDomain Text

tdrAutoRenew :: Lens' TransferDomain (Maybe Bool)

tdrAuthCode :: Lens' TransferDomain (Maybe Text)

tdrAdminContact :: Lens' TransferDomain ContactDetail

data TransferDomainResponse

Constructors

TransferDomainResponse

Fields

_tdsOperationId :: Text

Identifier for tracking the progress of the request. To use this ID to query the operation status, use GetOperationDetail. Type: String Default: None Constraints: Maximum 255 characters.

tdsOperationId :: Iso' TransferDomainResponse Text

data Route53Domains

Supported version (2014-05-15) of the Amazon Route 53 Domains service.

data ContactType

Indicates whether the contact is a person, company, association, or public organization. If you choose an option other than PERSON, you must enter an organization name, and you can't enable privacy protection for the contact. Type: String Default: None Constraints: Maximum 255 characters. Valid values: PERSON | COMPANY | ASSOCIATION | PUBLIC_BODY Parents: RegistrantContact, AdminContact, TechContact Required: Yes.

Constructors

ContactTypeAssociation
ASSOCIATION

ContactTypeCompany
COMPANY

ContactTypePerson
PERSON

ContactTypePublicBody
PUBLIC_BODY

ContactTypeReseller
RESELLER

data CountryCode

Code for the country of the contact's address. Type: String Default: None Constraints: Maximum 255 characters. Parents: RegistrantContact, AdminContact, TechContact Required: Yes.

Constructors

CountryCodeAd
AD

CountryCodeAe
AE

CountryCodeAf
AF

CountryCodeAg
AG

CountryCodeAi
AI

CountryCodeAl
AL

CountryCodeAm
AM

CountryCodeAn
AN

CountryCodeAo
AO

CountryCodeAq
AQ

CountryCodeAr
AR

CountryCodeAs
AS

CountryCodeAt
AT

CountryCodeAu
AU

CountryCodeAw
AW

CountryCodeAz
AZ

CountryCodeBa
BA

CountryCodeBb
BB

CountryCodeBd
BD

CountryCodeBe
BE

CountryCodeBf
BF

CountryCodeBg
BG

CountryCodeBh
BH

CountryCodeBi
BI

CountryCodeBj
BJ

CountryCodeBl
BL

CountryCodeBm
BM

CountryCodeBn
BN

CountryCodeBo
BO

CountryCodeBr
BR

CountryCodeBs
BS

CountryCodeBt
BT

CountryCodeBw
BW

CountryCodeBy
BY

CountryCodeBz
BZ

CountryCodeCa
CA

CountryCodeCc
CC

CountryCodeCd
CD

CountryCodeCf
CF

CountryCodeCg
CG

CountryCodeCh
CH

CountryCodeCi
CI

CountryCodeCk
CK

CountryCodeCl
CL

CountryCodeCm
CM

CountryCodeCn
CN

CountryCodeCo
CO

CountryCodeCr
CR

CountryCodeCu
CU

CountryCodeCv
CV

CountryCodeCx
CX

CountryCodeCy
CY

CountryCodeCz
CZ

CountryCodeDe
DE

CountryCodeDj
DJ

CountryCodeDk
DK

CountryCodeDm
DM

CountryCodeDo
DO

CountryCodeDz
DZ

CountryCodeEc
EC

CountryCodeEe
EE

CountryCodeEg
EG

CountryCodeEr
ER

CountryCodeEs
ES

CountryCodeEt
ET

CountryCodeFi
FI

CountryCodeFj
FJ

CountryCodeFk
FK

CountryCodeFm
FM

CountryCodeFo
FO

CountryCodeFr
FR

CountryCodeGa
GA

CountryCodeGb
GB

CountryCodeGd
GD

CountryCodeGe
GE

CountryCodeGh
GH

CountryCodeGi
GI

CountryCodeGl
GL

CountryCodeGm
GM

CountryCodeGn
GN

CountryCodeGq
GQ

CountryCodeGr
GR

CountryCodeGt
GT

CountryCodeGu
GU

CountryCodeGw
GW

CountryCodeGy
GY

CountryCodeHk
HK

CountryCodeHn
HN

CountryCodeHr
HR

CountryCodeHt
HT

CountryCodeHu
HU

CountryCodeId
ID

CountryCodeIe
IE

CountryCodeIl
IL

CountryCodeIm
IM

CountryCodeIn
IN

CountryCodeIq
IQ

CountryCodeIr
IR

CountryCodeIs
IS

CountryCodeIt
IT

CountryCodeJm
JM

CountryCodeJo
JO

CountryCodeJp
JP

CountryCodeKe
KE

CountryCodeKg
KG

CountryCodeKh
KH

CountryCodeKi
KI

CountryCodeKm
KM

CountryCodeKn
KN

CountryCodeKp
KP

CountryCodeKr
KR

CountryCodeKw
KW

CountryCodeKy
KY

CountryCodeKz
KZ

CountryCodeLa
LA

CountryCodeLb
LB

CountryCodeLc
LC

CountryCodeLi
LI

CountryCodeLk
LK

CountryCodeLr
LR

CountryCodeLs
LS

CountryCodeLt
LT

CountryCodeLu
LU

CountryCodeLv
LV

CountryCodeLy
LY

CountryCodeMa
MA

CountryCodeMc
MC

CountryCodeMd
MD

CountryCodeMe
ME

CountryCodeMf
MF

CountryCodeMg
MG

CountryCodeMh
MH

CountryCodeMk
MK

CountryCodeMl
ML

CountryCodeMm
MM

CountryCodeMn
MN

CountryCodeMo
MO

CountryCodeMp
MP

CountryCodeMr
MR

CountryCodeMs
MS

CountryCodeMt
MT

CountryCodeMu
MU

CountryCodeMv
MV

CountryCodeMw
MW

CountryCodeMx
MX

CountryCodeMy
MY

CountryCodeMz
MZ

CountryCodeNa
NA

CountryCodeNc
NC

CountryCodeNe
NE

CountryCodeNg
NG

CountryCodeNi
NI

CountryCodeNl
NL

CountryCodeNo
NO

CountryCodeNp
NP

CountryCodeNr
NR

CountryCodeNu
NU

CountryCodeNz
NZ

CountryCodeOm
OM

CountryCodePa
PA

CountryCodePe
PE

CountryCodePf
PF

CountryCodePg
PG

CountryCodePh
PH

CountryCodePk
PK

CountryCodePl
PL

CountryCodePm
PM

CountryCodePn
PN

CountryCodePr
PR

CountryCodePt
PT

CountryCodePw
PW

CountryCodePy
PY

CountryCodeQa
QA

CountryCodeRo
RO

CountryCodeRs
RS

CountryCodeRu
RU

CountryCodeRw
RW

CountryCodeSa
SA

CountryCodeSb
SB

CountryCodeSc
SC

CountryCodeSd
SD

CountryCodeSe
SE

CountryCodeSg
SG

CountryCodeSh
SH

CountryCodeSi
SI

CountryCodeSk
SK

CountryCodeSl
SL

CountryCodeSm
SM

CountryCodeSn
SN

CountryCodeSo
SO

CountryCodeSr
SR

CountryCodeSt
ST

CountryCodeSv
SV

CountryCodeSy
SY

CountryCodeSz
SZ

CountryCodeTc
TC

CountryCodeTd
TD

CountryCodeTg
TG

CountryCodeTh
TH

CountryCodeTj
TJ

CountryCodeTk
TK

CountryCodeTl
TL

CountryCodeTm
TM

CountryCodeTn
TN

CountryCodeTo
TO

CountryCodeTr
TR

CountryCodeTt
TT

CountryCodeTv
TV

CountryCodeTw
TW

CountryCodeTz
TZ

CountryCodeUa
UA

CountryCodeUg
UG

CountryCodeUs
US

CountryCodeUy
UY

CountryCodeUz
UZ

CountryCodeVa
VA

CountryCodeVc
VC

CountryCodeVe
VE

CountryCodeVg
VG

CountryCodeVi
VI

CountryCodeVn
VN

CountryCodeVu
VU

CountryCodeWf
WF

CountryCodeWs
WS

CountryCodeYe
YE

CountryCodeYt
YT

CountryCodeZa
ZA

CountryCodeZm
ZM

CountryCodeZw
ZW

data DomainAvailability

Whether the domain name is available for registering. You can only register domains designated as AVAILABLE. Type: String Valid values: AVAILABLE – The domain name is available. AVAILABLE_RESERVED – The domain name is reserved under specific conditions. AVAILABLE_PREORDER – The domain name is available and can be preordered. UNAVAILABLE – The domain name is not available. UNAVAILABLE_PREMIUM – The domain name is not available. UNAVAILABLE_RESTRICTED – The domain name is forbidden. RESERVED – The domain name has been reserved for another person or organization.

Constructors

DomainAvailabilityAvailable
AVAILABLE

DomainAvailabilityAvailablePreorder
AVAILABLE_PREORDER

DomainAvailabilityAvailableReserved
AVAILABLE_RESERVED

DomainAvailabilityReserved
RESERVED

DomainAvailabilityUnavailable
UNAVAILABLE

DomainAvailabilityUnavailablePremium
UNAVAILABLE_PREMIUM

DomainAvailabilityUnavailableRestricted
UNAVAILABLE_RESTRICTED

data ExtraParamName

Name of the additional parameter required by the top-level domain. Type: String Default: None Valid values: DUNS_NUMBER | BRAND_NUMBER | BIRTH_DEPARTMENT | BIRTH_DATE_IN_YYYY_MM_DD | BIRTH_COUNTRY | BIRTH_CITY | DOCUMENT_NUMBER | AU_ID_NUMBER | AU_ID_TYPE | CA_LEGAL_TYPE | FI_BUSINESS_NUMBER | FI_ID_NUMBER | IT_PIN | RU_PASSPORT_DATA | SE_ID_NUMBER | SG_ID_NUMBER | VAT_NUMBER Parent: ExtraParams Required: Yes.

Constructors

ExtraParamNameAuIdNumber
AU_ID_NUMBER

ExtraParamNameAuIdType
AU_ID_TYPE

ExtraParamNameBirthCity
BIRTH_CITY

ExtraParamNameBirthCountry
BIRTH_COUNTRY

ExtraParamNameBirthDateInYyyyMmDd
BIRTH_DATE_IN_YYYY_MM_DD

ExtraParamNameBirthDepartment
BIRTH_DEPARTMENT

ExtraParamNameBrandNumber
BRAND_NUMBER

ExtraParamNameCaLegalType
CA_LEGAL_TYPE

ExtraParamNameDocumentNumber
DOCUMENT_NUMBER

ExtraParamNameDunsNumber
DUNS_NUMBER

ExtraParamNameFiBusinessNumber
FI_BUSINESS_NUMBER

ExtraParamNameFiIdNumber
FI_ID_NUMBER

ExtraParamNameItPin
IT_PIN

ExtraParamNameRuPassportData
RU_PASSPORT_DATA

ExtraParamNameSeIdNumber
SE_ID_NUMBER

ExtraParamNameSgIdNumber
SG_ID_NUMBER

ExtraParamNameVatNumber
VAT_NUMBER

data OperationStatus

The current status of the requested operation in the system. Type: String.

Constructors

OperationStatusError
ERROR

OperationStatusFailed
FAILED

OperationStatusInProgress
IN_PROGRESS

OperationStatusSubmitted
SUBMITTED

OperationStatusSuccessful
SUCCESSFUL

data OperationType

Type of the action requested. Type: String Valid values: REGISTER_DOMAIN | DELETE_DOMAIN | TRANSFER_IN_DOMAIN | UPDATE_DOMAIN_CONTACT | UPDATE_NAMESERVER | CHANGE_PRIVACY_PROTECTION | DOMAIN_LOCK.

Constructors

OperationTypeChangePrivacyProtection
CHANGE_PRIVACY_PROTECTION

OperationTypeDeleteDomain
DELETE_DOMAIN

OperationTypeDomainLock
DOMAIN_LOCK

OperationTypeRegisterDomain
REGISTER_DOMAIN

OperationTypeTransferInDomain
TRANSFER_IN_DOMAIN

OperationTypeUpdateDomainContact
UPDATE_DOMAIN_CONTACT

OperationTypeUpdateNameserver
UPDATE_NAMESERVER

data ContactDetail

Provides details about the domain registrant. Type: Complex Children: FirstName, MiddleName, LastName, ContactType, OrganizationName, AddressLine1, AddressLine2, City, State, CountryCode, ZipCode, PhoneNumber, Email, Fax, ExtraParams.

Constructors

ContactDetail

Fields

_cdOrganizationName :: Maybe Text

Name of the organization for contact types other than PERSON. Type: String Default: None Constraints: Maximum 255 characters. Contact type must not be PERSON. Parents: RegistrantContact, AdminContact, TechContact Required: No.

_cdEmail :: Maybe Text

Email address of the contact. Type: String Default: None Constraints: Maximum 254 characters. Parents: RegistrantContact, AdminContact, TechContact Required: Yes.

_cdState :: Maybe Text

The state or province of the contact's city. Type: String Default: None Constraints: Maximum 255 characters. Parents: RegistrantContact, AdminContact, TechContact Required: No.

_cdFax :: Maybe Text

Fax number of the contact. Type: String Default: None Constraints: Phone number must be specified in the format "+[country dialing code].[number including any area code]". For example, a US phone number might appear as "+1.1234567890". Parents: RegistrantContact, AdminContact, TechContact Required: No.

_cdLastName :: Maybe Text

Last name of contact. Type: String Default: None Constraints: Maximum 255 characters. Parents: RegistrantContact, AdminContact, TechContact Required: Yes.

_cdExtraParams :: [ExtraParam]

A list of name-value pairs for parameters required by certain top-level domains. Type: Complex Default: None Parents: RegistrantContact, AdminContact, TechContact Children: Name, Value Required: No.

_cdZipCode :: Maybe Text

The zip or postal code of the contact's address. Type: String Default: None Constraints: Maximum 255 characters. Parents: RegistrantContact, AdminContact, TechContact Required: No.

_cdAddressLine1 :: Maybe Text

First line of the contact's address. Type: String Default: None Constraints: Maximum 255 characters. Parents: RegistrantContact, AdminContact, TechContact Required: Yes.

_cdCity :: Maybe Text

The city of the contact's address. Type: String Default: None Constraints: Maximum 255 characters. Parents: RegistrantContact, AdminContact, TechContact Required: Yes.

_cdPhoneNumber :: Maybe Text

The phone number of the contact. Type: String Default: None Constraints: Phone number must be specified in the format "+[country dialing code].[number including any area code&gt;]". For example, a US phone number might appear as "+1.1234567890". Parents: RegistrantContact, AdminContact, TechContact Required: Yes.

_cdAddressLine2 :: Maybe Text

Second line of contact's address, if any. Type: String Default: None Constraints: Maximum 255 characters. Parents: RegistrantContact, AdminContact, TechContact Required: No.

_cdFirstName :: Maybe Text

First name of contact. Type: String Default: None Constraints: Maximum 255 characters. Parents: RegistrantContact, AdminContact, TechContact Required: Yes.

_cdCountryCode :: Maybe CountryCode

Code for the country of the contact's address. Type: String Default: None Constraints: Maximum 255 characters. Parents: RegistrantContact, AdminContact, TechContact Required: Yes.

_cdContactType :: Maybe ContactType

Indicates whether the contact is a person, company, association, or public organization. If you choose an option other than PERSON, you must enter an organization name, and you can't enable privacy protection for the contact. Type: String Default: None Constraints: Maximum 255 characters. Valid values: PERSON | COMPANY | ASSOCIATION | PUBLIC_BODY Parents: RegistrantContact, AdminContact, TechContact Required: Yes.

data DomainSummary

Constructors

DomainSummary

Fields

_ddyExpiry :: Maybe ISO8601

Expiration date of the domain in Coordinated Universal Time (UTC). Type: Long.

_ddyTransferLock :: Maybe Bool

Indicates whether a domain is locked from unauthorized transfer to another party. Type: Boolean Valid values: True | False.

_ddyAutoRenew :: Maybe Bool

Indicates whether the domain is automatically renewed upon expiration. Type: Boolean Valid values: True | False.

_ddyDomainName :: Text

The name of a domain. Type: String.

data ExtraParam

ExtraParam includes the following elements.

Constructors

ExtraParam

Fields

_epValue :: Text

Values corresponding to the additional parameter names required by some top-level domains. Type: String Default: None Constraints: Maximum 2048 characters. Parent: ExtraParams Required: Yes.

_epName :: ExtraParamName

Name of the additional parameter required by the top-level domain. Type: String Default: None Valid values: DUNS_NUMBER | BRAND_NUMBER | BIRTH_DEPARTMENT | BIRTH_DATE_IN_YYYY_MM_DD | BIRTH_COUNTRY | BIRTH_CITY | DOCUMENT_NUMBER | AU_ID_NUMBER | AU_ID_TYPE | CA_LEGAL_TYPE | FI_BUSINESS_NUMBER | FI_ID_NUMBER | IT_PIN | RU_PASSPORT_DATA | SE_ID_NUMBER | SG_ID_NUMBER | VAT_NUMBER Parent: ExtraParams Required: Yes.

data Nameserver

Nameserver includes the following elements.

Constructors

Nameserver

Fields

_nName :: Text

The fully qualified host name of the name server. Type: String Constraint: Maximum 255 characterss Parent: Nameservers.

_nGlueIps :: [Text]

Glue IP address of a name server entry. Glue IP addresses are required only when the name of the name server is a subdomain of the domain. For example, if your domain is example.com and the name server for the domain is ns.example.com, you need to specify the IP address for ns.example.com. Type: List of IP addresses. Constraints: The list can contain only one IPv4 and one IPv6 address. Parent: Nameservers.

data OperationSummary

OperationSummary includes the following elements.

Constructors

OperationSummary

Fields

_osStatus :: OperationStatus

The current status of the requested operation in the system. Type: String.

_osSubmittedDate :: ISO8601

The date when the request was submitted.

_osOperationId :: Text

Identifier returned to track the requested action. Type: String.

_osType :: OperationType

Type of the action requested. Type: String Valid values: REGISTER_DOMAIN | DELETE_DOMAIN | TRANSFER_IN_DOMAIN | UPDATE_DOMAIN_CONTACT | UPDATE_NAMESERVER | CHANGE_PRIVACY_PROTECTION | DOMAIN_LOCK.

cdZipCode :: Lens' ContactDetail (Maybe Text)

cdState :: Lens' ContactDetail (Maybe Text)

cdPhoneNumber :: Lens' ContactDetail (Maybe Text)

cdOrganizationName :: Lens' ContactDetail (Maybe Text)

cdLastName :: Lens' ContactDetail (Maybe Text)

cdFirstName :: Lens' ContactDetail (Maybe Text)

cdFax :: Lens' ContactDetail (Maybe Text)

cdExtraParams :: Lens' ContactDetail [ExtraParam]

cdEmail :: Lens' ContactDetail (Maybe Text)

cdCountryCode :: Lens' ContactDetail (Maybe CountryCode)

cdContactType :: Lens' ContactDetail (Maybe ContactType)

cdCity :: Lens' ContactDetail (Maybe Text)

cdAddressLine2 :: Lens' ContactDetail (Maybe Text)

cdAddressLine1 :: Lens' ContactDetail (Maybe Text)

ddyTransferLock :: Lens' DomainSummary (Maybe Bool)

ddyExpiry :: Lens' DomainSummary (Maybe ISO8601)

ddyDomainName :: Lens' DomainSummary Text

ddyAutoRenew :: Lens' DomainSummary (Maybe Bool)

epValue :: Lens' ExtraParam Text

epName :: Lens' ExtraParam ExtraParamName

nName :: Lens' Nameserver Text

nGlueIps :: Lens' Nameserver [Text]

osType :: Lens' OperationSummary OperationType

osSubmittedDate :: Lens' OperationSummary ISO8601

osStatus :: Lens' OperationSummary OperationStatus

osOperationId :: Lens' OperationSummary Text

updateDomainContact

Arguments

:: Text
_udcrDomainName

-> UpdateDomainContact

Minimum specification for a UpdateDomainContact request.

data UpdateDomainContact

Constructors

UpdateDomainContact

Fields

_udcrDomainName :: Text

The name of a domain. Type: String Default: None Constraints: The domain name can contain only the letters a through z, the numbers 0 through 9, and hyphen (-). Internationalized Domain Names are not supported. Required: Yes.

_udcrRegistrantContact :: Maybe ContactDetail

Provides detailed contact information. Type: Complex Children: FirstName, MiddleName, LastName, ContactType, OrganizationName, AddressLine1, AddressLine2, City, State, CountryCode, ZipCode, PhoneNumber, Email, Fax, ExtraParams Required: Yes.

_udcrAdminContact :: Maybe ContactDetail

Provides detailed contact information. Type: Complex Children: FirstName, MiddleName, LastName, ContactType, OrganizationName, AddressLine1, AddressLine2, City, State, CountryCode, ZipCode, PhoneNumber, Email, Fax, ExtraParams Required: Yes.

_udcrTechContact :: Maybe ContactDetail

Provides detailed contact information. Type: Complex Children: FirstName, MiddleName, LastName, ContactType, OrganizationName, AddressLine1, AddressLine2, City, State, CountryCode, ZipCode, PhoneNumber, Email, Fax, ExtraParams Required: Yes.

udcrTechContact :: Lens' UpdateDomainContact (Maybe ContactDetail)

udcrRegistrantContact :: Lens' UpdateDomainContact (Maybe ContactDetail)

udcrDomainName :: Lens' UpdateDomainContact Text

udcrAdminContact :: Lens' UpdateDomainContact (Maybe ContactDetail)

data UpdateDomainContactResponse

Constructors

UpdateDomainContactResponse

Fields

_udcsOperationId :: Text

Identifier for tracking the progress of the request. To use this ID to query the operation status, use GetOperationDetail. Type: String Default: None Constraints: Maximum 255 characters.

udcsOperationId :: Iso' UpdateDomainContactResponse Text

updateDomainContactPrivacy

Arguments

:: Text
_udcprDomainName

-> UpdateDomainContactPrivacy

Minimum specification for a UpdateDomainContactPrivacy request.

data UpdateDomainContactPrivacy

Constructors

UpdateDomainContactPrivacy

Fields

_udcprDomainName :: Text

The name of a domain. Type: String Default: None Constraints: The domain name can contain only the letters a through z, the numbers 0 through 9, and hyphen (-). Internationalized Domain Names are not supported. Required: Yes.

_udcprTechPrivacy :: Maybe Bool

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter. Type: Boolean Default: None Valid values: true | false Required: No.

_udcprRegistrantPrivacy :: Maybe Bool

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter. Type: Boolean Default: None Valid values: true | false Required: No.

_udcprAdminPrivacy :: Maybe Bool

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter. Type: Boolean Default: None Valid values: true | false Required: No.

udcprTechPrivacy :: Lens' UpdateDomainContactPrivacy (Maybe Bool)

udcprRegistrantPrivacy :: Lens' UpdateDomainContactPrivacy (Maybe Bool)

udcprDomainName :: Lens' UpdateDomainContactPrivacy Text

udcprAdminPrivacy :: Lens' UpdateDomainContactPrivacy (Maybe Bool)

data UpdateDomainContactPrivacyResponse

Constructors

UpdateDomainContactPrivacyResponse

Fields

_udcpsOperationId :: Text

Identifier for tracking the progress of the request. To use this ID to query the operation status, use GetOperationDetail. Type: String Default: None Constraints: Maximum 255 characters.

udcpsOperationId :: Iso' UpdateDomainContactPrivacyResponse Text

data UpdateDomainNameservers

Constructors

UpdateDomainNameservers

Fields

_udnrDomainName :: Text

The name of a domain. Type: String Default: None Constraints: The domain name can contain only the letters a through z, the numbers 0 through 9, and hyphen (-). Internationalized Domain Names are not supported. Required: Yes.

_udnrNameservers :: [Nameserver]

A list of new name servers for the domain. Type: Complex Children: Name, GlueIps Required: Yes.

udnrNameservers :: Lens' UpdateDomainNameservers [Nameserver]

udnrDomainName :: Lens' UpdateDomainNameservers Text

data UpdateDomainNameserversResponse

Constructors

UpdateDomainNameserversResponse

Fields

_udnsOperationId :: Text

Identifier for tracking the progress of the request. To use this ID to query the operation status, use GetOperationDetail. Type: String Default: None Constraints: Maximum 255 characters.

udnsOperationId :: Iso' UpdateDomainNameserversResponse Text