Network.AWS.Route53Domains.V2014_05_15.EnableDomainTransferLock

Description

This operation sets the transfer lock on the domain (specifically the clientTransferProhibited status) to prevent domain transfers. Successful submission returns an operation ID that you can use to track the progress and completion of the action. If the request is not completed successfully, the domain registrant will be notified by email. EnableDomainTransferLock Example POST HTTP1.1 host:route53domains.us-east-1.amazonaws.com x-amz-date:20140711T205230Z authorization:AWS4-HMAC-SHA256 Credential=AKIAIOSFODNN7EXAMPLE20140711us-east-1route53domainsaws4_request,

SignedHeaders=content-length;content-type;host;user-agent;x-amz-date;x-amz-target, Signature=[calculated-signature] x-amz-target:Route53Domains_v20140515.EnableDomainTransferLock user-agent:aws-sdk-java1.8.3 Linux2.6.18-164.el5PAE Java_HotSpot (TM )_Server_VM24.60-b091.7.0_60 content-type:application/x-amz-json-1.1 content-length:[number of characters in the JSON string] { DomainName:"example.com" } HTTP/1.1 200 Content-Length:[number of characters in the JSON string] { OperationId:"0b370c79-faa4-40fe-94c8-b423069de3f6" }.

Documentation

data EnableDomainTransferLock

Constructors

EnableDomainTransferLock

Fields

_edtlrDomainName :: Text

The name of a domain. Type: String Default: None Constraints: The domain name can contain only the letters a through z, the numbers 0 through 9, and hyphen (-). Internationalized Domain Names are not supported. Required: Yes.

edtlrDomainName :: Iso' EnableDomainTransferLock Text

data EnableDomainTransferLockResponse

Constructors

EnableDomainTransferLockResponse

Fields

_edtlsOperationId :: Text

Identifier for tracking the progress of the request. To use this ID to query the operation status, use GetOperationDetail. Type: String Default: None Constraints: Maximum 255 characters.

edtlsOperationId :: Iso' EnableDomainTransferLockResponse Text