Network.AWS.Route53Domains.V2014_05_15.GetDomainDetail

Description

This operation returns detailed information about the domain. The domain's contact information is also returned as part of the output. GetDomainDetail Example POST HTTP1.1 host:route53domains.us-east-1.amazonaws.com x-amz-date:20140711T205230Z authorization:AWS4-HMAC-SHA256 Credential=AKIAIOSFODNN7EXAMPLE20140711us-east-1route53domainsaws4_request,

SignedHeaders=content-length;content-type;host;user-agent;x-amz-date;x-amz-target, Signature=[calculated-signature] x-amz-target:Route53Domains_v20140515.GetDomainDetail user-agent:aws-sdk-java1.8.3 Linux2.6.18-164.el5PAE Java_HotSpot (TM )_Server_VM24.60-b091.7.0_60 content-type:application/x-amz-json-1.1 content-length:[number of characters in the JSON string] { DomainName:"example.com" } HTTP/1.1 200 Content-Length:[number of characters in the JSON string] { AbuseContactEmail:"abusesupport.gandi.net", AbuseContactPhone:"+33.170377661", AdminContact:{ AddressLine1:"1 Any Street", AddressLine2:"", City:Anytown, CountryCode:US, Email:"johnexample.com", ExtraParams:[ ], FirstName:John, LastName:Doe, PhoneNumber:"+2065550100", State:WA, ZipCode:"98101" }, AdminPrivacy:true, AutoRenew:true, CreationDate:1400010459, DomainName:"example.com", ExpirationDate:1431539259, Nameservers:[ { GlueIps:[ ], Name:"ns-2048.awsdns-64.com" }, { GlueIps:[ ], Name:"ns-2051.awsdns-67.co.uk" }, { GlueIps:[ ], Name:"ns-2050.awsdns-66.org" }, { GlueIps:[ ], Name:"ns-2049.awsdns-65.net" } ], RegistrantContact:{ AddressLine1:"1 Any Street", AddressLine2:"", City:Anytown, CountryCode:US, Email:"johnexample.com", ExtraParams:[ ], FirstName:John, LastName:Doe, PhoneNumber:"+2065550100", State:WA, ZipCode:"98101" }, RegistrantPrivacy:true, RegistrarName:"GANDI SAS", RegistrarUrl:"http://www.gandi.net", Reseller:Amazon, StatusList:[ "clientTransferProhibited" ], TechContact:{ AddressLine1:"1 Any Street", AddressLine2:"", City:Anytown, CountryCode:US, Email:"johnexample.com", ExtraParams:[ ], FirstName:John, LastName:Doe, PhoneNumber:"+2065550100", State:WA, ZipCode:"98101" }, TechPrivacy:true, UpdatedDate:1400010459, WhoIsServer:"whois.gandi.net" }.

Request

data GetDomainDetail

The GetDomainDetail request includes the following element.

Request constructor

mkGetDomainDetail

Arguments

:: Text
gddDomainName

-> GetDomainDetail

Smart constructor for the minimum required parameters to construct a valid GetDomainDetail request.

Request lenses

gddDomainName :: Lens' GetDomainDetail Text

The name of a domain. Type: String Default: None Constraints: The domain name can contain only the letters a through z, the numbers 0 through 9, and hyphen (-). Internationalized Domain Names are not supported. Required: Yes.

Response

data GetDomainDetailResponse

The GetDomainDetail response includes the following elements.

Response lenses

gddrsDomainName :: Lens' GetDomainDetailResponse Text

The name of a domain. Type: String.

gddrsNameservers :: Lens' GetDomainDetailResponse [Nameserver]

The name of the domain. Type: String.

gddrsAutoRenew :: Lens' GetDomainDetailResponse (Maybe Bool)

Specifies whether the domain registration is set to renew automatically. Type: Boolean.

gddrsAdminContact :: Lens' GetDomainDetailResponse ContactDetail

Provides details about the domain administrative contact. Type: Complex Children: FirstName, MiddleName, LastName, ContactType, OrganizationName, AddressLine1, AddressLine2, City, State, CountryCode, ZipCode, PhoneNumber, Email, Fax, ExtraParams.

gddrsRegistrantContact :: Lens' GetDomainDetailResponse ContactDetail

Provides details about the domain registrant. Type: Complex Children: FirstName, MiddleName, LastName, ContactType, OrganizationName, AddressLine1, AddressLine2, City, State, CountryCode, ZipCode, PhoneNumber, Email, Fax, ExtraParams.

gddrsTechContact :: Lens' GetDomainDetailResponse ContactDetail

Provides details about the domain technical contact. Type: Complex Children: FirstName, MiddleName, LastName, ContactType, OrganizationName, AddressLine1, AddressLine2, City, State, CountryCode, ZipCode, PhoneNumber, Email, Fax, ExtraParams.

gddrsAdminPrivacy :: Lens' GetDomainDetailResponse (Maybe Bool)

Specifies whether contact information for the admin contact is concealed from WHOIS queries. If the value is true, WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter. Type: Boolean.

gddrsRegistrantPrivacy :: Lens' GetDomainDetailResponse (Maybe Bool)

Specifies whether contact information for the registrant contact is concealed from WHOIS queries. If the value is true, WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter. Type: Boolean.

gddrsTechPrivacy :: Lens' GetDomainDetailResponse (Maybe Bool)

Specifies whether contact information for the tech contact is concealed from WHOIS queries. If the value is true, WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter. Type: Boolean.

gddrsRegistrarName :: Lens' GetDomainDetailResponse (Maybe Text)

Name of the registrar of the domain as identified in the registry. Amazon Route 53 domains are registered by registrar Gandi. The value is "GANDI SAS". Type: String.

gddrsWhoIsServer :: Lens' GetDomainDetailResponse (Maybe Text)

The fully qualified name of the WHOIS server that can answer the WHOIS query for the domain. Type: String.

gddrsRegistrarUrl :: Lens' GetDomainDetailResponse (Maybe Text)

Web address of the registrar. Type: String.

gddrsAbuseContactEmail :: Lens' GetDomainDetailResponse (Maybe Text)

Email address to contact to report incorrect contact information for a domain, to report that the domain is being used to send spam, to report that someone is cybersquatting on a domain name, or report some other type of abuse. Type: String.

gddrsAbuseContactPhone :: Lens' GetDomainDetailResponse (Maybe Text)

Phone number for reporting abuse. Type: String.

gddrsRegistryDomainId :: Lens' GetDomainDetailResponse (Maybe Text)

Reserved for future use.

gddrsCreationDate :: Lens' GetDomainDetailResponse (Maybe ISO8601)

The date when the domain was created as found in the response to a WHOIS query. The date format is Unix time.

gddrsUpdatedDate :: Lens' GetDomainDetailResponse (Maybe ISO8601)

The last updated date of the domain as found in the response to a WHOIS query. The date format is Unix time.

gddrsExpirationDate :: Lens' GetDomainDetailResponse (Maybe ISO8601)

The date when the registration for the domain is set to expire. The date format is Unix time.

gddrsReseller :: Lens' GetDomainDetailResponse (Maybe Text)

Reseller of the domain. Domains registered or transferred using Amazon Route 53 domains will have Amazon as the reseller. Type: String.

gddrsDnsSec :: Lens' GetDomainDetailResponse (Maybe Text)

Reserved for future use.

gddrsStatusList :: Lens' GetDomainDetailResponse [Text]

An array of domain name status codes, also known as Extensible Provisioning Protocol (EPP) status codes. ICANN, the organization that maintains a central database of domain names, has developed a set of domain name status codes that tell you the status of a variety of operations on a domain name, for example, registering a domain name, transferring a domain name to another registrar, renewing the registration for a domain name, and so on. All registrars use this same set of status codes. For a current list of domain name status codes and an explanation of what each code means, go to the ICANN website and search for epp status codes. (Search on the ICANN website; web searches sometimes return an old version of the document.) Type: Array of String.