Network.AWS.Route53Domains.V2014_05_15.GetOperationDetail

Description

This operation returns the current status of an operation that is not completed. GetOperationDetail Example POST HTTP1.1 host:route53domains.us-east-1.amazonaws.com x-amz-date:20140711T205230Z authorization:AWS4-HMAC-SHA256 Credential=AKIAIOSFODNN7EXAMPLE20140711us-east-1route53domainsaws4_request,

SignedHeaders=content-length;content-type;host;user-agent;x-amz-date;x-amz-target, Signature=[calculated-signature] x-amz-target:Route53Domains_v20140515.GetOperationDetail user-agent:aws-sdk-java1.8.3 Linux2.6.18-164.el5PAE Java_HotSpot (TM )_Server_VM24.60-b091.7.0_60 content-type:application/x-amz-json-1.1 content-length:[number of characters in the JSON string] { OperationId:"43884ce5-e30a-4801-858f-7aa86356c127" } HTTP/1.1 200 Content-Length:[number of characters in the JSON string] { DomainName:"happierdomain.ca", OperationId:"43884ce5-e30a-4801-858f-7aa86356c127", Status:WORKFLOW_IN_PROGRESS, SubmittedDate : 1402630939.057, Type : REGISTER_DOMAIN }.

Documentation

data GetOperationDetail

Constructors

GetOperationDetail

Fields

_godrOperationId :: Text

The identifier for the operation for which you want to get the status. Amazon Route 53 returned the identifier in the response to the original request. Type: String Default: None Required: Yes.

godrOperationId :: Iso' GetOperationDetail Text

data GetOperationDetailResponse

Constructors

GetOperationDetailResponse

Fields

_godsDomainName :: Maybe Text

The name of a domain. Type: String.

_godsMessage :: Maybe Text

Detailed information on the status including possible errors. Type: String.

_godsOperationId :: Maybe Text

The identifier for the operation. Type: String.

_godsStatus :: Maybe OperationStatus

The current status of the requested operation in the system. Type: String.

_godsType :: Maybe OperationType

The type of operation that was requested. Type: String.

_godsSubmittedDate :: Maybe ISO8601

The date when the request was submitted.

godsType :: Lens' GetOperationDetailResponse (Maybe OperationType)

godsSubmittedDate :: Lens' GetOperationDetailResponse (Maybe ISO8601)

godsStatus :: Lens' GetOperationDetailResponse (Maybe OperationStatus)

godsOperationId :: Lens' GetOperationDetailResponse (Maybe Text)

godsMessage :: Lens' GetOperationDetailResponse (Maybe Text)

godsDomainName :: Lens' GetOperationDetailResponse (Maybe Text)