Network.AWS.Route53Domains.V2014_05_15.ListDomains

Description

This operation returns all the domain names registered with Amazon Route 53 for the current AWS account. ListDomains Example POST HTTP1.1 host:route53domains.us-east-1.amazonaws.com x-amz-date:20140711T205230Z authorization:AWS4-HMAC-SHA256 Credential=AKIAIOSFODNN7EXAMPLE20140711us-east-1route53domainsaws4_request,

SignedHeaders=content-length;content-type;host;user-agent;x-amz-date;x-amz-target, Signature=[calculated-signature] x-amz-target:Route53Domains_v20140515.ListDomains user-agent:aws-sdk-java1.8.3 Linux2.6.18-164.el5PAE Java_HotSpot (TM )_Server_VM24.60-b091.7.0_60 content-type:application/x-amz-json-1.1 content-length:[number of characters in the JSON string] { Marker:"AxDAClaROQAXasf29GHWAIKPLA=", MaxItems:20 } HTTP/1.1 200 Content-Length:[number of characters in the JSON string] { Domains:[ { AutoRenew:false, DomainName:"example.com", Expiry:1431203765, TransferLock:false }, { AutoRenew:false, DomainName:"example.net", Expiry:1431539260, TransferLock:false }, { AutoRenew:false, DomainName:"example.org", Expiry:1431240024, TransferLock:false }, { AutoRenew:false, DomainName:"example.test", Expiry:1431539259, TransferLock:false } ] }.

Request

data ListDomains

The ListDomains request includes the following elements.

Request constructor

mkListDomains :: ListDomains

Smart constructor for the minimum required parameters to construct a valid ListDomains request.

Request lenses

ldMarker :: Lens' ListDomains (Maybe Text)

For an initial request for a list of domains, omit this element. If the number of domains that are associated with the current AWS account is greater than the value that you specified for MaxItems, you can use Marker to return additional domains. Get the value of NextPageMarker from the previous response, and submit another request that includes the value of NextPageMarker in the Marker element. Type: String Default: None Constraints: The marker must match the value specified in the previous request. Required: No.

ldMaxItems :: Lens' ListDomains (Maybe Integer)

Number of domains to be returned. Type: Integer Default: 20 Constraints: A numeral between 1 and 100. Required: No.

Response

data ListDomainsResponse

The ListDomains response includes the following elements.

Response lenses

ldrsDomains :: Lens' ListDomainsResponse [DomainSummary]

A summary of domains. Type: Complex type containing a list of domain summaries. Children: AutoRenew, DomainName, Expiry, TransferLock.

ldrsNextPageMarker :: Lens' ListDomainsResponse (Maybe Text)

If there are more domains than you specified for MaxItems in the request, submit another request and include the value of NextPageMarker in the value of Marker. Type: String Parent: Operations.