Network.AWS.Route53Domains.V2014_05_15.Types

Description

Pending.

Documentation

data Route53Domains

Supported version (2014-05-15) of the Amazon Route 53 Domains service.

data ContactType

Indicates whether the contact is a person, company, association, or public organization. If you choose an option other than PERSON, you must enter an organization name, and you can't enable privacy protection for the contact. Type: String Default: None Constraints: Maximum 255 characters. Valid values: PERSON | COMPANY | ASSOCIATION | PUBLIC_BODY Parents: RegistrantContact, AdminContact, TechContact Required: Yes.

Constructors

ContactTypeAssociation
ASSOCIATION

ContactTypeCompany
COMPANY

ContactTypePerson
PERSON

ContactTypePublicBody
PUBLIC_BODY

ContactTypeReseller
RESELLER

data CountryCode

Code for the country of the contact's address. Type: String Default: None Constraints: Maximum 255 characters. Parents: RegistrantContact, AdminContact, TechContact Required: Yes.

Constructors

CountryCodeAd
AD

CountryCodeAe
AE

CountryCodeAf
AF

CountryCodeAg
AG

CountryCodeAi
AI

CountryCodeAl
AL

CountryCodeAm
AM

CountryCodeAn
AN

CountryCodeAo
AO

CountryCodeAq
AQ

CountryCodeAr
AR

CountryCodeAs
AS

CountryCodeAt
AT

CountryCodeAu
AU

CountryCodeAw
AW

CountryCodeAz
AZ

CountryCodeBa
BA

CountryCodeBb
BB

CountryCodeBd
BD

CountryCodeBe
BE

CountryCodeBf
BF

CountryCodeBg
BG

CountryCodeBh
BH

CountryCodeBi
BI

CountryCodeBj
BJ

CountryCodeBl
BL

CountryCodeBm
BM

CountryCodeBn
BN

CountryCodeBo
BO

CountryCodeBr
BR

CountryCodeBs
BS

CountryCodeBt
BT

CountryCodeBw
BW

CountryCodeBy
BY

CountryCodeBz
BZ

CountryCodeCa
CA

CountryCodeCc
CC

CountryCodeCd
CD

CountryCodeCf
CF

CountryCodeCg
CG

CountryCodeCh
CH

CountryCodeCi
CI

CountryCodeCk
CK

CountryCodeCl
CL

CountryCodeCm
CM

CountryCodeCn
CN

CountryCodeCo
CO

CountryCodeCr
CR

CountryCodeCu
CU

CountryCodeCv
CV

CountryCodeCx
CX

CountryCodeCy
CY

CountryCodeCz
CZ

CountryCodeDe
DE

CountryCodeDj
DJ

CountryCodeDk
DK

CountryCodeDm
DM

CountryCodeDo
DO

CountryCodeDz
DZ

CountryCodeEc
EC

CountryCodeEe
EE

CountryCodeEg
EG

CountryCodeEr
ER

CountryCodeEs
ES

CountryCodeEt
ET

CountryCodeFi
FI

CountryCodeFj
FJ

CountryCodeFk
FK

CountryCodeFm
FM

CountryCodeFo
FO

CountryCodeFr
FR

CountryCodeGa
GA

CountryCodeGb
GB

CountryCodeGd
GD

CountryCodeGe
GE

CountryCodeGh
GH

CountryCodeGi
GI

CountryCodeGl
GL

CountryCodeGm
GM

CountryCodeGn
GN

CountryCodeGq
GQ

CountryCodeGr
GR

CountryCodeGt
GT

CountryCodeGu
GU

CountryCodeGw
GW

CountryCodeGy
GY

CountryCodeHk
HK

CountryCodeHn
HN

CountryCodeHr
HR

CountryCodeHt
HT

CountryCodeHu
HU

CountryCodeId
ID

CountryCodeIe
IE

CountryCodeIl
IL

CountryCodeIm
IM

CountryCodeIn
IN

CountryCodeIq
IQ

CountryCodeIr
IR

CountryCodeIs
IS

CountryCodeIt
IT

CountryCodeJm
JM

CountryCodeJo
JO

CountryCodeJp
JP

CountryCodeKe
KE

CountryCodeKg
KG

CountryCodeKh
KH

CountryCodeKi
KI

CountryCodeKm
KM

CountryCodeKn
KN

CountryCodeKp
KP

CountryCodeKr
KR

CountryCodeKw
KW

CountryCodeKy
KY

CountryCodeKz
KZ

CountryCodeLa
LA

CountryCodeLb
LB

CountryCodeLc
LC

CountryCodeLi
LI

CountryCodeLk
LK

CountryCodeLr
LR

CountryCodeLs
LS

CountryCodeLt
LT

CountryCodeLu
LU

CountryCodeLv
LV

CountryCodeLy
LY

CountryCodeMa
MA

CountryCodeMc
MC

CountryCodeMd
MD

CountryCodeMe
ME

CountryCodeMf
MF

CountryCodeMg
MG

CountryCodeMh
MH

CountryCodeMk
MK

CountryCodeMl
ML

CountryCodeMm
MM

CountryCodeMn
MN

CountryCodeMo
MO

CountryCodeMp
MP

CountryCodeMr
MR

CountryCodeMs
MS

CountryCodeMt
MT

CountryCodeMu
MU

CountryCodeMv
MV

CountryCodeMw
MW

CountryCodeMx
MX

CountryCodeMy
MY

CountryCodeMz
MZ

CountryCodeNa
NA

CountryCodeNc
NC

CountryCodeNe
NE

CountryCodeNg
NG

CountryCodeNi
NI

CountryCodeNl
NL

CountryCodeNo
NO

CountryCodeNp
NP

CountryCodeNr
NR

CountryCodeNu
NU

CountryCodeNz
NZ

CountryCodeOm
OM

CountryCodePa
PA

CountryCodePe
PE

CountryCodePf
PF

CountryCodePg
PG

CountryCodePh
PH

CountryCodePk
PK

CountryCodePl
PL

CountryCodePm
PM

CountryCodePn
PN

CountryCodePr
PR

CountryCodePt
PT

CountryCodePw
PW

CountryCodePy
PY

CountryCodeQa
QA

CountryCodeRo
RO

CountryCodeRs
RS

CountryCodeRu
RU

CountryCodeRw
RW

CountryCodeSa
SA

CountryCodeSb
SB

CountryCodeSc
SC

CountryCodeSd
SD

CountryCodeSe
SE

CountryCodeSg
SG

CountryCodeSh
SH

CountryCodeSi
SI

CountryCodeSk
SK

CountryCodeSl
SL

CountryCodeSm
SM

CountryCodeSn
SN

CountryCodeSo
SO

CountryCodeSr
SR

CountryCodeSt
ST

CountryCodeSv
SV

CountryCodeSy
SY

CountryCodeSz
SZ

CountryCodeTc
TC

CountryCodeTd
TD

CountryCodeTg
TG

CountryCodeTh
TH

CountryCodeTj
TJ

CountryCodeTk
TK

CountryCodeTl
TL

CountryCodeTm
TM

CountryCodeTn
TN

CountryCodeTo
TO

CountryCodeTr
TR

CountryCodeTt
TT

CountryCodeTv
TV

CountryCodeTw
TW

CountryCodeTz
TZ

CountryCodeUa
UA

CountryCodeUg
UG

CountryCodeUs
US

CountryCodeUy
UY

CountryCodeUz
UZ

CountryCodeVa
VA

CountryCodeVc
VC

CountryCodeVe
VE

CountryCodeVg
VG

CountryCodeVi
VI

CountryCodeVn
VN

CountryCodeVu
VU

CountryCodeWf
WF

CountryCodeWs
WS

CountryCodeYe
YE

CountryCodeYt
YT

CountryCodeZa
ZA

CountryCodeZm
ZM

CountryCodeZw
ZW

data DomainAvailability

Whether the domain name is available for registering. You can only register domains designated as AVAILABLE. Type: String Valid values: AVAILABLE – The domain name is available. AVAILABLE_RESERVED – The domain name is reserved under specific conditions. AVAILABLE_PREORDER – The domain name is available and can be preordered. UNAVAILABLE – The domain name is not available. UNAVAILABLE_PREMIUM – The domain name is not available. UNAVAILABLE_RESTRICTED – The domain name is forbidden. RESERVED – The domain name has been reserved for another person or organization.

Constructors

DomainAvailabilityAvailable
AVAILABLE

DomainAvailabilityAvailablePreorder
AVAILABLE_PREORDER

DomainAvailabilityAvailableReserved
AVAILABLE_RESERVED

DomainAvailabilityReserved
RESERVED

DomainAvailabilityUnavailable
UNAVAILABLE

DomainAvailabilityUnavailablePremium
UNAVAILABLE_PREMIUM

DomainAvailabilityUnavailableRestricted
UNAVAILABLE_RESTRICTED

data ExtraParamName

Name of the additional parameter required by the top-level domain. Type: String Default: None Valid values: DUNS_NUMBER | BRAND_NUMBER | BIRTH_DEPARTMENT | BIRTH_DATE_IN_YYYY_MM_DD | BIRTH_COUNTRY | BIRTH_CITY | DOCUMENT_NUMBER | AU_ID_NUMBER | AU_ID_TYPE | CA_LEGAL_TYPE | FI_BUSINESS_NUMBER | FI_ID_NUMBER | IT_PIN | RU_PASSPORT_DATA | SE_ID_NUMBER | SG_ID_NUMBER | VAT_NUMBER Parent: ExtraParams Required: Yes.

Constructors

ExtraParamNameAuIdNumber
AU_ID_NUMBER

ExtraParamNameAuIdType
AU_ID_TYPE

ExtraParamNameBirthCity
BIRTH_CITY

ExtraParamNameBirthCountry
BIRTH_COUNTRY

ExtraParamNameBirthDateInYyyyMmDd
BIRTH_DATE_IN_YYYY_MM_DD

ExtraParamNameBirthDepartment
BIRTH_DEPARTMENT

ExtraParamNameBrandNumber
BRAND_NUMBER

ExtraParamNameCaLegalType
CA_LEGAL_TYPE

ExtraParamNameDocumentNumber
DOCUMENT_NUMBER

ExtraParamNameDunsNumber
DUNS_NUMBER

ExtraParamNameFiBusinessNumber
FI_BUSINESS_NUMBER

ExtraParamNameFiIdNumber
FI_ID_NUMBER

ExtraParamNameItPin
IT_PIN

ExtraParamNameRuPassportData
RU_PASSPORT_DATA

ExtraParamNameSeIdNumber
SE_ID_NUMBER

ExtraParamNameSgIdNumber
SG_ID_NUMBER

ExtraParamNameVatNumber
VAT_NUMBER

data OperationStatus

The current status of the requested operation in the system. Type: String.

Constructors

OperationStatusError
ERROR

OperationStatusFailed
FAILED

OperationStatusInProgress
IN_PROGRESS

OperationStatusSubmitted
SUBMITTED

OperationStatusSuccessful
SUCCESSFUL

data OperationType

Type of the action requested. Type: String Valid values: REGISTER_DOMAIN | DELETE_DOMAIN | TRANSFER_IN_DOMAIN | UPDATE_DOMAIN_CONTACT | UPDATE_NAMESERVER | CHANGE_PRIVACY_PROTECTION | DOMAIN_LOCK.

Constructors

OperationTypeChangePrivacyProtection
CHANGE_PRIVACY_PROTECTION

OperationTypeDeleteDomain
DELETE_DOMAIN

OperationTypeDomainLock
DOMAIN_LOCK

OperationTypeRegisterDomain
REGISTER_DOMAIN

OperationTypeTransferInDomain
TRANSFER_IN_DOMAIN

OperationTypeUpdateDomainContact
UPDATE_DOMAIN_CONTACT

OperationTypeUpdateNameserver
UPDATE_NAMESERVER

data ContactDetail

Provides details about the domain registrant. Type: Complex Children: FirstName, MiddleName, LastName, ContactType, OrganizationName, AddressLine1, AddressLine2, City, State, CountryCode, ZipCode, PhoneNumber, Email, Fax, ExtraParams.

Constructors

ContactDetail

Fields

_cdOrganizationName :: Maybe Text

Name of the organization for contact types other than PERSON. Type: String Default: None Constraints: Maximum 255 characters. Contact type must not be PERSON. Parents: RegistrantContact, AdminContact, TechContact Required: No.

_cdEmail :: Maybe Text

Email address of the contact. Type: String Default: None Constraints: Maximum 254 characters. Parents: RegistrantContact, AdminContact, TechContact Required: Yes.

_cdState :: Maybe Text

The state or province of the contact's city. Type: String Default: None Constraints: Maximum 255 characters. Parents: RegistrantContact, AdminContact, TechContact Required: No.

_cdFax :: Maybe Text

Fax number of the contact. Type: String Default: None Constraints: Phone number must be specified in the format "+[country dialing code].[number including any area code]". For example, a US phone number might appear as "+1.1234567890". Parents: RegistrantContact, AdminContact, TechContact Required: No.

_cdLastName :: Maybe Text

Last name of contact. Type: String Default: None Constraints: Maximum 255 characters. Parents: RegistrantContact, AdminContact, TechContact Required: Yes.

_cdExtraParams :: [ExtraParam]

A list of name-value pairs for parameters required by certain top-level domains. Type: Complex Default: None Parents: RegistrantContact, AdminContact, TechContact Children: Name, Value Required: No.

_cdZipCode :: Maybe Text

The zip or postal code of the contact's address. Type: String Default: None Constraints: Maximum 255 characters. Parents: RegistrantContact, AdminContact, TechContact Required: No.

_cdAddressLine1 :: Maybe Text

First line of the contact's address. Type: String Default: None Constraints: Maximum 255 characters. Parents: RegistrantContact, AdminContact, TechContact Required: Yes.

_cdCity :: Maybe Text

The city of the contact's address. Type: String Default: None Constraints: Maximum 255 characters. Parents: RegistrantContact, AdminContact, TechContact Required: Yes.

_cdPhoneNumber :: Maybe Text

The phone number of the contact. Type: String Default: None Constraints: Phone number must be specified in the format "+[country dialing code].[number including any area code&gt;]". For example, a US phone number might appear as "+1.1234567890". Parents: RegistrantContact, AdminContact, TechContact Required: Yes.

_cdAddressLine2 :: Maybe Text

Second line of contact's address, if any. Type: String Default: None Constraints: Maximum 255 characters. Parents: RegistrantContact, AdminContact, TechContact Required: No.

_cdFirstName :: Maybe Text

First name of contact. Type: String Default: None Constraints: Maximum 255 characters. Parents: RegistrantContact, AdminContact, TechContact Required: Yes.

_cdCountryCode :: Maybe CountryCode

Code for the country of the contact's address. Type: String Default: None Constraints: Maximum 255 characters. Parents: RegistrantContact, AdminContact, TechContact Required: Yes.

_cdContactType :: Maybe ContactType

Indicates whether the contact is a person, company, association, or public organization. If you choose an option other than PERSON, you must enter an organization name, and you can't enable privacy protection for the contact. Type: String Default: None Constraints: Maximum 255 characters. Valid values: PERSON | COMPANY | ASSOCIATION | PUBLIC_BODY Parents: RegistrantContact, AdminContact, TechContact Required: Yes.

data DomainSummary

Constructors

DomainSummary

Fields

_ddyExpiry :: Maybe ISO8601

Expiration date of the domain in Coordinated Universal Time (UTC). Type: Long.

_ddyTransferLock :: Maybe Bool

Indicates whether a domain is locked from unauthorized transfer to another party. Type: Boolean Valid values: True | False.

_ddyAutoRenew :: Maybe Bool

Indicates whether the domain is automatically renewed upon expiration. Type: Boolean Valid values: True | False.

_ddyDomainName :: Text

The name of a domain. Type: String.

data ExtraParam

ExtraParam includes the following elements.

Constructors

ExtraParam

Fields

_epValue :: Text

Values corresponding to the additional parameter names required by some top-level domains. Type: String Default: None Constraints: Maximum 2048 characters. Parent: ExtraParams Required: Yes.

_epName :: ExtraParamName

Name of the additional parameter required by the top-level domain. Type: String Default: None Valid values: DUNS_NUMBER | BRAND_NUMBER | BIRTH_DEPARTMENT | BIRTH_DATE_IN_YYYY_MM_DD | BIRTH_COUNTRY | BIRTH_CITY | DOCUMENT_NUMBER | AU_ID_NUMBER | AU_ID_TYPE | CA_LEGAL_TYPE | FI_BUSINESS_NUMBER | FI_ID_NUMBER | IT_PIN | RU_PASSPORT_DATA | SE_ID_NUMBER | SG_ID_NUMBER | VAT_NUMBER Parent: ExtraParams Required: Yes.

data Nameserver

Nameserver includes the following elements.

Constructors

Nameserver

Fields

_nName :: Text

The fully qualified host name of the name server. Type: String Constraint: Maximum 255 characterss Parent: Nameservers.

_nGlueIps :: [Text]

Glue IP address of a name server entry. Glue IP addresses are required only when the name of the name server is a subdomain of the domain. For example, if your domain is example.com and the name server for the domain is ns.example.com, you need to specify the IP address for ns.example.com. Type: List of IP addresses. Constraints: The list can contain only one IPv4 and one IPv6 address. Parent: Nameservers.

data OperationSummary

OperationSummary includes the following elements.

Constructors

OperationSummary

Fields

_osStatus :: OperationStatus

The current status of the requested operation in the system. Type: String.

_osSubmittedDate :: ISO8601

The date when the request was submitted.

_osOperationId :: Text

Identifier returned to track the requested action. Type: String.

_osType :: OperationType

Type of the action requested. Type: String Valid values: REGISTER_DOMAIN | DELETE_DOMAIN | TRANSFER_IN_DOMAIN | UPDATE_DOMAIN_CONTACT | UPDATE_NAMESERVER | CHANGE_PRIVACY_PROTECTION | DOMAIN_LOCK.

cdZipCode :: Lens' ContactDetail (Maybe Text)

cdState :: Lens' ContactDetail (Maybe Text)

cdPhoneNumber :: Lens' ContactDetail (Maybe Text)

cdOrganizationName :: Lens' ContactDetail (Maybe Text)

cdLastName :: Lens' ContactDetail (Maybe Text)

cdFirstName :: Lens' ContactDetail (Maybe Text)

cdFax :: Lens' ContactDetail (Maybe Text)

cdExtraParams :: Lens' ContactDetail [ExtraParam]

cdEmail :: Lens' ContactDetail (Maybe Text)

cdCountryCode :: Lens' ContactDetail (Maybe CountryCode)

cdContactType :: Lens' ContactDetail (Maybe ContactType)

cdCity :: Lens' ContactDetail (Maybe Text)

cdAddressLine2 :: Lens' ContactDetail (Maybe Text)

cdAddressLine1 :: Lens' ContactDetail (Maybe Text)

ddyTransferLock :: Lens' DomainSummary (Maybe Bool)

ddyExpiry :: Lens' DomainSummary (Maybe ISO8601)

ddyDomainName :: Lens' DomainSummary Text

ddyAutoRenew :: Lens' DomainSummary (Maybe Bool)

epValue :: Lens' ExtraParam Text

epName :: Lens' ExtraParam ExtraParamName

nName :: Lens' Nameserver Text

nGlueIps :: Lens' Nameserver [Text]

osType :: Lens' OperationSummary OperationType

osSubmittedDate :: Lens' OperationSummary ISO8601

osStatus :: Lens' OperationSummary OperationStatus

osOperationId :: Lens' OperationSummary Text