Network.AWS.Route53Domains.GetDomainDetail

Description

This operation returns detailed information about a specified domain that is associated with the current AWS account. Contact information for the domain is also returned as part of the output.

Creating a Request

getDomainDetail

Arguments

:: Text
gddDomainName

-> GetDomainDetail

Creates a value of GetDomainDetail with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gddDomainName - The name of the domain that you want to get detailed information about.

data GetDomainDetail

The GetDomainDetail request includes the following element.

See: getDomainDetail smart constructor.

Request Lenses

gddDomainName :: Lens' GetDomainDetail Text

The name of the domain that you want to get detailed information about.

Destructuring the Response

getDomainDetailResponse

Arguments

:: Int
gddrsResponseStatus

-> Text
gddrsDomainName

-> ContactDetail
gddrsAdminContact

-> ContactDetail
gddrsRegistrantContact

-> ContactDetail
gddrsTechContact

-> GetDomainDetailResponse

Creates a value of GetDomainDetailResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gddrsTechPrivacy - Specifies whether contact information for the tech contact is concealed from WHOIS queries. If the value is true , WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter.

gddrsDNSSec - Reserved for future use.

gddrsWhoIsServer - The fully qualified name of the WHOIS server that can answer the WHOIS query for the domain.

gddrsRegistryDomainId - Reserved for future use.

gddrsRegistrantPrivacy - Specifies whether contact information for the registrant contact is concealed from WHOIS queries. If the value is true , WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter.

gddrsUpdatedDate - The last updated date of the domain as found in the response to a WHOIS query. The date format is Unix time.

gddrsAdminPrivacy - Specifies whether contact information for the admin contact is concealed from WHOIS queries. If the value is true , WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter.

gddrsAutoRenew - Specifies whether the domain registration is set to renew automatically.

gddrsAbuseContactPhone - Phone number for reporting abuse.

gddrsRegistrarURL - Web address of the registrar.

gddrsAbuseContactEmail - Email address to contact to report incorrect contact information for a domain, to report that the domain is being used to send spam, to report that someone is cybersquatting on a domain name, or report some other type of abuse.

gddrsExpirationDate - The date when the registration for the domain is set to expire. The date format is Unix time.

gddrsCreationDate - The date when the domain was created as found in the response to a WHOIS query. The date format is Unix time.

gddrsRegistrarName - Name of the registrar of the domain as identified in the registry. Amazon Route 53 domains are registered by registrar Gandi. The value is "GANDI SAS" .

gddrsReseller - Reseller of the domain. Domains registered or transferred using Amazon Route 53 domains will have Amazon as the reseller.

gddrsStatusList - An array of domain name status codes, also known as Extensible Provisioning Protocol (EPP) status codes. ICANN, the organization that maintains a central database of domain names, has developed a set of domain name status codes that tell you the status of a variety of operations on a domain name, for example, registering a domain name, transferring a domain name to another registrar, renewing the registration for a domain name, and so on. All registrars use this same set of status codes. For a current list of domain name status codes and an explanation of what each code means, go to the ICANN website and search for epp status codes . (Search on the ICANN website; web searches sometimes return an old version of the document.)

gddrsResponseStatus - -- | The response status code.

gddrsDomainName - The name of a domain.

gddrsNameservers - The name of the domain.

gddrsAdminContact - Provides details about the domain administrative contact.

gddrsRegistrantContact - Provides details about the domain registrant.

gddrsTechContact - Provides details about the domain technical contact.

data GetDomainDetailResponse

The GetDomainDetail response includes the following elements.

See: getDomainDetailResponse smart constructor.

Response Lenses

gddrsTechPrivacy :: Lens' GetDomainDetailResponse (Maybe Bool)

Specifies whether contact information for the tech contact is concealed from WHOIS queries. If the value is true , WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter.

gddrsDNSSec :: Lens' GetDomainDetailResponse (Maybe Text)

Reserved for future use.

gddrsWhoIsServer :: Lens' GetDomainDetailResponse (Maybe Text)

The fully qualified name of the WHOIS server that can answer the WHOIS query for the domain.

gddrsRegistryDomainId :: Lens' GetDomainDetailResponse (Maybe Text)

Reserved for future use.

gddrsRegistrantPrivacy :: Lens' GetDomainDetailResponse (Maybe Bool)

Specifies whether contact information for the registrant contact is concealed from WHOIS queries. If the value is true , WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter.

gddrsUpdatedDate :: Lens' GetDomainDetailResponse (Maybe UTCTime)

The last updated date of the domain as found in the response to a WHOIS query. The date format is Unix time.

gddrsAdminPrivacy :: Lens' GetDomainDetailResponse (Maybe Bool)

Specifies whether contact information for the admin contact is concealed from WHOIS queries. If the value is true , WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter.

gddrsAutoRenew :: Lens' GetDomainDetailResponse (Maybe Bool)

Specifies whether the domain registration is set to renew automatically.

gddrsAbuseContactPhone :: Lens' GetDomainDetailResponse (Maybe Text)

Phone number for reporting abuse.

gddrsRegistrarURL :: Lens' GetDomainDetailResponse (Maybe Text)

Web address of the registrar.

gddrsAbuseContactEmail :: Lens' GetDomainDetailResponse (Maybe Text)

Email address to contact to report incorrect contact information for a domain, to report that the domain is being used to send spam, to report that someone is cybersquatting on a domain name, or report some other type of abuse.

gddrsExpirationDate :: Lens' GetDomainDetailResponse (Maybe UTCTime)

The date when the registration for the domain is set to expire. The date format is Unix time.

gddrsCreationDate :: Lens' GetDomainDetailResponse (Maybe UTCTime)

The date when the domain was created as found in the response to a WHOIS query. The date format is Unix time.

gddrsRegistrarName :: Lens' GetDomainDetailResponse (Maybe Text)

Name of the registrar of the domain as identified in the registry. Amazon Route 53 domains are registered by registrar Gandi. The value is "GANDI SAS" .

gddrsReseller :: Lens' GetDomainDetailResponse (Maybe Text)

Reseller of the domain. Domains registered or transferred using Amazon Route 53 domains will have Amazon as the reseller.

gddrsStatusList :: Lens' GetDomainDetailResponse [Text]

An array of domain name status codes, also known as Extensible Provisioning Protocol (EPP) status codes. ICANN, the organization that maintains a central database of domain names, has developed a set of domain name status codes that tell you the status of a variety of operations on a domain name, for example, registering a domain name, transferring a domain name to another registrar, renewing the registration for a domain name, and so on. All registrars use this same set of status codes. For a current list of domain name status codes and an explanation of what each code means, go to the ICANN website and search for epp status codes . (Search on the ICANN website; web searches sometimes return an old version of the document.)

gddrsResponseStatus :: Lens' GetDomainDetailResponse Int

- | The response status code.

gddrsDomainName :: Lens' GetDomainDetailResponse Text

The name of a domain.

gddrsNameservers :: Lens' GetDomainDetailResponse [Nameserver]

The name of the domain.

gddrsAdminContact :: Lens' GetDomainDetailResponse ContactDetail

Provides details about the domain administrative contact.

gddrsRegistrantContact :: Lens' GetDomainDetailResponse ContactDetail

Provides details about the domain registrant.

gddrsTechContact :: Lens' GetDomainDetailResponse ContactDetail

Provides details about the domain technical contact.