Network.AWS.Route53Domains.ListDomains

Description

This operation returns all the domain names registered with Amazon Route 53 for the current AWS account.

This operation returns paginated results.

Creating a Request

listDomains :: ListDomains

Creates a value of ListDomains with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ldMarker - For an initial request for a list of domains, omit this element. If the number of domains that are associated with the current AWS account is greater than the value that you specified for MaxItems , you can use Marker to return additional domains. Get the value of NextPageMarker from the previous response, and submit another request that includes the value of NextPageMarker in the Marker element. Constraints: The marker must match the value specified in the previous request.

ldMaxItems - Number of domains to be returned. Default: 20

data ListDomains

The ListDomains request includes the following elements.

See: listDomains smart constructor.

Request Lenses

ldMarker :: Lens' ListDomains (Maybe Text)

For an initial request for a list of domains, omit this element. If the number of domains that are associated with the current AWS account is greater than the value that you specified for MaxItems , you can use Marker to return additional domains. Get the value of NextPageMarker from the previous response, and submit another request that includes the value of NextPageMarker in the Marker element. Constraints: The marker must match the value specified in the previous request.

ldMaxItems :: Lens' ListDomains (Maybe Int)

Number of domains to be returned. Default: 20

Destructuring the Response

listDomainsResponse

Arguments

:: Int
ldrsResponseStatus

-> ListDomainsResponse

Creates a value of ListDomainsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ldrsNextPageMarker - If there are more domains than you specified for MaxItems in the request, submit another request and include the value of NextPageMarker in the value of Marker .

ldrsResponseStatus - -- | The response status code.

ldrsDomains - A summary of domains.

data ListDomainsResponse

The ListDomains response includes the following elements.

See: listDomainsResponse smart constructor.

Response Lenses

ldrsNextPageMarker :: Lens' ListDomainsResponse (Maybe Text)

If there are more domains than you specified for MaxItems in the request, submit another request and include the value of NextPageMarker in the value of Marker .

ldrsResponseStatus :: Lens' ListDomainsResponse Int

- | The response status code.

ldrsDomains :: Lens' ListDomainsResponse [DomainSummary]

A summary of domains.