Network.AWS.Route53Domains.ListOperations

Description

This operation returns the operation IDs of operations that are not yet complete.

http://docs.aws.amazon.com/Route53/latest/APIReference/api-ListOperations.html

Request

data ListOperations

Request constructor

listOperations :: ListOperations

ListOperations constructor.

The fields accessible through corresponding lenses are:

loMarker :: Maybe Text

loMaxItems :: Maybe Int

Request lenses

loMarker :: Lens' ListOperations (Maybe Text)

For an initial request for a list of operations, omit this element. If the number of operations that are not yet complete is greater than the value that you specified for MaxItems, you can use Marker to return additional operations. Get the value of NextPageMarker from the previous response, and submit another request that includes the value of NextPageMarker in the Marker element.

Type: String

Default: None

Required: No

loMaxItems :: Lens' ListOperations (Maybe Int)

Number of domains to be returned.

Type: Integer

Default: 20

Constraints: A value between 1 and 100.

Required: No

Response

data ListOperationsResponse

Response constructor

listOperationsResponse :: ListOperationsResponse

ListOperationsResponse constructor.

The fields accessible through corresponding lenses are:

lorNextPageMarker :: Maybe Text

lorOperations :: [OperationSummary]

Response lenses

lorNextPageMarker :: Lens' ListOperationsResponse (Maybe Text)

If there are more operations than you specified for MaxItems in the request, submit another request and include the value of NextPageMarker in the value of Marker.

Type: String

Parent: Operations

lorOperations :: Lens' ListOperationsResponse [OperationSummary]

Lists summaries of the operations.

Type: Complex type containing a list of operation summaries

Children: OperationId, Status, SubmittedDate, Type