Network.AWS.Route53Domains.RegisterDomain

Description

This operation registers a domain. Domains are registered by the AWS registrar partner, Gandi. For some top-level domains (TLDs), this operation requires extra parameters.

When you register a domain, Amazon Route 53 does the following:

Creates a Amazon Route 53 hosted zone that has the same name as the domain. Amazon Route 53 assigns four name servers to your hosted zone and automatically updates your domain registration with the names of these name servers.

Enables autorenew, so your domain registration will renew automatically each year. We'll notify you in advance of the renewal date so you can choose whether to renew the registration.

Optionally enables privacy protection, so WHOIS queries return contact information for our registrar partner, Gandi, instead of the information you entered for registrant, admin, and tech contacts.

If registration is successful, returns an operation ID that you can use to track the progress and completion of the action. If the request is not completed successfully, the domain registrant is notified by email.

Charges your AWS account an amount based on the top-level domain. For more information, see Amazon Route 53 Pricing .

Creating a Request

registerDomain

Arguments

:: Text
rDomainName

-> Natural
rDurationInYears

-> ContactDetail
rAdminContact

-> ContactDetail
rRegistrantContact

-> ContactDetail
rTechContact

-> RegisterDomain

Creates a value of RegisterDomain with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rPrivacyProtectTechContact - Whether you want to conceal contact information from WHOIS queries. If you specify true , WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter. Default: true

rPrivacyProtectRegistrantContact - Whether you want to conceal contact information from WHOIS queries. If you specify true , WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter. Default: true

rAutoRenew - Indicates whether the domain will be automatically renewed (true ) or not (false ). Autorenewal only takes effect after the account is charged. Default: true

rPrivacyProtectAdminContact - Whether you want to conceal contact information from WHOIS queries. If you specify true , WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter. Default: true

rIdNLangCode - Reserved for future use.

rDomainName - The domain name that you want to register. Constraints: The domain name can contain only the letters a through z, the numbers 0 through 9, and hyphen (-). Internationalized Domain Names are not supported.

rDurationInYears - The number of years that you want to register the domain for. Domains are registered for a minimum of one year. The maximum period depends on the top-level domain. For the range of valid values for your domain, see Domains that You Can Register with Amazon Route 53 in the Amazon Route 53 Developer Guide . Default: 1

rAdminContact - Provides detailed contact information.

rRegistrantContact - Provides detailed contact information.

rTechContact - Provides detailed contact information.

data RegisterDomain

The RegisterDomain request includes the following elements.

See: registerDomain smart constructor.

Request Lenses

rPrivacyProtectTechContact :: Lens' RegisterDomain (Maybe Bool)

Whether you want to conceal contact information from WHOIS queries. If you specify true , WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter. Default: true

rPrivacyProtectRegistrantContact :: Lens' RegisterDomain (Maybe Bool)

Whether you want to conceal contact information from WHOIS queries. If you specify true , WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter. Default: true

rAutoRenew :: Lens' RegisterDomain (Maybe Bool)

Indicates whether the domain will be automatically renewed (true ) or not (false ). Autorenewal only takes effect after the account is charged. Default: true

rPrivacyProtectAdminContact :: Lens' RegisterDomain (Maybe Bool)

Whether you want to conceal contact information from WHOIS queries. If you specify true , WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter. Default: true

rIdNLangCode :: Lens' RegisterDomain (Maybe Text)

Reserved for future use.

rDomainName :: Lens' RegisterDomain Text

The domain name that you want to register. Constraints: The domain name can contain only the letters a through z, the numbers 0 through 9, and hyphen (-). Internationalized Domain Names are not supported.

rDurationInYears :: Lens' RegisterDomain Natural

The number of years that you want to register the domain for. Domains are registered for a minimum of one year. The maximum period depends on the top-level domain. For the range of valid values for your domain, see Domains that You Can Register with Amazon Route 53 in the Amazon Route 53 Developer Guide . Default: 1

rAdminContact :: Lens' RegisterDomain ContactDetail

Provides detailed contact information.

rRegistrantContact :: Lens' RegisterDomain ContactDetail

Provides detailed contact information.

rTechContact :: Lens' RegisterDomain ContactDetail

Provides detailed contact information.

Destructuring the Response

registerDomainResponse

Arguments

:: Int
rdrsResponseStatus

-> Text
rdrsOperationId

-> RegisterDomainResponse

Creates a value of RegisterDomainResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rdrsResponseStatus - -- | The response status code.

rdrsOperationId - Identifier for tracking the progress of the request. To use this ID to query the operation status, use GetOperationDetail .

data RegisterDomainResponse

The RegisterDomain response includes the following element.

See: registerDomainResponse smart constructor.

Response Lenses

rdrsResponseStatus :: Lens' RegisterDomainResponse Int

- | The response status code.

rdrsOperationId :: Lens' RegisterDomainResponse Text

Identifier for tracking the progress of the request. To use this ID to query the operation status, use GetOperationDetail .