Network.AWS.Route53Domains.RetrieveDomainAuthCode

Description

This operation returns the AuthCode for the domain. To transfer a domain to another registrar, you provide this value to the new registrar.

http://docs.aws.amazon.com/Route53/latest/APIReference/api-RetrieveDomainAuthCode.html

Request

data RetrieveDomainAuthCode

Request constructor

retrieveDomainAuthCode

Arguments

:: Text
rdacDomainName

-> RetrieveDomainAuthCode

RetrieveDomainAuthCode constructor.

The fields accessible through corresponding lenses are:

rdacDomainName :: Text

Request lenses

rdacDomainName :: Lens' RetrieveDomainAuthCode Text

The name of a domain.

Type: String

Default: None

Constraints: The domain name can contain only the letters a through z, the numbers 0 through 9, and hyphen (-). Internationalized Domain Names are not supported.

Required: Yes

Response

data RetrieveDomainAuthCodeResponse

Response constructor

retrieveDomainAuthCodeResponse

Arguments

:: Text
rdacrAuthCode

-> RetrieveDomainAuthCodeResponse

RetrieveDomainAuthCodeResponse constructor.

The fields accessible through corresponding lenses are:

rdacrAuthCode :: Text

Response lenses

rdacrAuthCode :: Lens' RetrieveDomainAuthCodeResponse Text

The authorization code for the domain.

Type: String