Network.AWS.Route53Domains.TransferDomain

Description

This operation transfers a domain from another registrar to Amazon Route 53. When the transfer is complete, the domain is registered with the AWS registrar partner, Gandi.

For transfer requirements, a detailed procedure, and information about viewing the status of a domain transfer, see Transferring Registration for a Domain to Amazon Route 53 in the Amazon Route 53 Developer Guide .

If the registrar for your domain is also the DNS service provider for the domain, we highly recommend that you consider transferring your DNS service to Amazon Route 53 or to another DNS service provider before you transfer your registration. Some registrars provide free DNS service when you purchase a domain registration. When you transfer the registration, the previous registrar will not renew your domain registration and could end your DNS service at any time.

Important: If the registrar for your domain is also the DNS service provider for the domain and you don't transfer DNS service to another provider, your website, email, and the web applications associated with the domain might become unavailable.

If the transfer is successful, this method returns an operation ID that you can use to track the progress and completion of the action. If the transfer doesn't complete successfully, the domain registrant will be notified by email.

Creating a Request

transferDomain

Arguments

:: Text
tdDomainName

-> Natural
tdDurationInYears

-> ContactDetail
tdAdminContact

-> ContactDetail
tdRegistrantContact

-> ContactDetail
tdTechContact

-> TransferDomain

Creates a value of TransferDomain with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tdPrivacyProtectTechContact - Whether you want to conceal contact information from WHOIS queries. If you specify true , WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter. Default: true

tdPrivacyProtectRegistrantContact - Whether you want to conceal contact information from WHOIS queries. If you specify true , WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter. Default: true

tdAutoRenew - Indicates whether the domain will be automatically renewed (true) or not (false). Autorenewal only takes effect after the account is charged. Default: true

tdPrivacyProtectAdminContact - Whether you want to conceal contact information from WHOIS queries. If you specify true , WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter. Default: true

tdIdNLangCode - Reserved for future use.

tdAuthCode - The authorization code for the domain. You get this value from the current registrar.

tdNameservers - Contains details for the host and glue IP addresses.

tdDomainName - The name of the domain that you want to transfer to Amazon Route 53. Constraints: The domain name can contain only the letters a through z, the numbers 0 through 9, and hyphen (-). Internationalized Domain Names are not supported.

tdDurationInYears - The number of years that you want to register the domain for. Domains are registered for a minimum of one year. The maximum period depends on the top-level domain. Default: 1

tdAdminContact - Provides detailed contact information.

tdRegistrantContact - Provides detailed contact information.

tdTechContact - Provides detailed contact information.

data TransferDomain

The TransferDomain request includes the following elements.

See: transferDomain smart constructor.

Request Lenses

tdPrivacyProtectTechContact :: Lens' TransferDomain (Maybe Bool)

Whether you want to conceal contact information from WHOIS queries. If you specify true , WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter. Default: true

tdPrivacyProtectRegistrantContact :: Lens' TransferDomain (Maybe Bool)

Whether you want to conceal contact information from WHOIS queries. If you specify true , WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter. Default: true

tdAutoRenew :: Lens' TransferDomain (Maybe Bool)

Indicates whether the domain will be automatically renewed (true) or not (false). Autorenewal only takes effect after the account is charged. Default: true

tdPrivacyProtectAdminContact :: Lens' TransferDomain (Maybe Bool)

Whether you want to conceal contact information from WHOIS queries. If you specify true , WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter. Default: true

tdIdNLangCode :: Lens' TransferDomain (Maybe Text)

Reserved for future use.

tdAuthCode :: Lens' TransferDomain (Maybe Text)

The authorization code for the domain. You get this value from the current registrar.

tdNameservers :: Lens' TransferDomain [Nameserver]

Contains details for the host and glue IP addresses.

tdDomainName :: Lens' TransferDomain Text

The name of the domain that you want to transfer to Amazon Route 53. Constraints: The domain name can contain only the letters a through z, the numbers 0 through 9, and hyphen (-). Internationalized Domain Names are not supported.

tdDurationInYears :: Lens' TransferDomain Natural

The number of years that you want to register the domain for. Domains are registered for a minimum of one year. The maximum period depends on the top-level domain. Default: 1

tdAdminContact :: Lens' TransferDomain ContactDetail

Provides detailed contact information.

tdRegistrantContact :: Lens' TransferDomain ContactDetail

Provides detailed contact information.

tdTechContact :: Lens' TransferDomain ContactDetail

Provides detailed contact information.

Destructuring the Response

transferDomainResponse

Arguments

:: Int
tdrsResponseStatus

-> Text
tdrsOperationId

-> TransferDomainResponse

Creates a value of TransferDomainResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tdrsResponseStatus - -- | The response status code.

tdrsOperationId - Identifier for tracking the progress of the request. To use this ID to query the operation status, use GetOperationDetail .

data TransferDomainResponse

The TranserDomain response includes the following element.

See: transferDomainResponse smart constructor.

Response Lenses

tdrsResponseStatus :: Lens' TransferDomainResponse Int

- | The response status code.

tdrsOperationId :: Lens' TransferDomainResponse Text

Identifier for tracking the progress of the request. To use this ID to query the operation status, use GetOperationDetail .