Network.AWS.Route53Domains.TransferDomain

Description

This operation transfers a domain from another registrar to Amazon Route 53. Domains are registered by the AWS registrar, Gandi upon transfer.

To transfer a domain, you need to meet all the domain transfer criteria, including the following:

You must supply nameservers to transfer a domain. You must disable the domain transfer lock (if any) before transferring the domain. A minimum of 60 days must have elapsed since the domain's registration or last transfer. We recommend you use the Amazon Route 53 as the DNS service for your domain. You can create a hosted zone in Amazon Route 53 for your current domain before transferring your domain.

Note that upon transfer, the domain duration is extended for a year if not otherwise specified. Autorenew is enabled by default.

If the transfer is successful, this method returns an operation ID that you can use to track the progress and completion of the action. If the request is not completed successfully, the domain registrant will be notified by email.

Transferring domains charges your AWS account an amount based on the top-level domain. For more information, see Amazon Route 53 Pricing.

http://docs.aws.amazon.com/Route53/latest/APIReference/api-TransferDomain.html

Request

data TransferDomain

Request constructor

transferDomain

Arguments

:: Text
tdDomainName

-> Natural
tdDurationInYears

-> ContactDetail
tdAdminContact

-> ContactDetail
tdRegistrantContact

-> ContactDetail
tdTechContact

-> TransferDomain

TransferDomain constructor.

The fields accessible through corresponding lenses are:

tdAdminContact :: ContactDetail

tdAuthCode :: Maybe Text

tdAutoRenew :: Maybe Bool

tdDomainName :: Text

tdDurationInYears :: Natural

tdIdnLangCode :: Maybe Text

tdNameservers :: [Nameserver]

tdPrivacyProtectAdminContact :: Maybe Bool

tdPrivacyProtectRegistrantContact :: Maybe Bool

tdPrivacyProtectTechContact :: Maybe Bool

tdRegistrantContact :: ContactDetail

tdTechContact :: ContactDetail

Request lenses

tdAdminContact :: Lens' TransferDomain ContactDetail

Provides detailed contact information.

Type: Complex

Children: FirstName, MiddleName, LastName, ContactType, OrganizationName, AddressLine1, AddressLine2, City, State, CountryCode, ZipCode, PhoneNumber, Email, Fax, ExtraParams

Required: Yes

tdAuthCode :: Lens' TransferDomain (Maybe Text)

The authorization code for the domain. You get this value from the current registrar.

Type: String

Required: Yes

tdAutoRenew :: Lens' TransferDomain (Maybe Bool)

Indicates whether the domain will be automatically renewed (true) or not (false). Autorenewal only takes effect after the account is charged.

Type: Boolean

Valid values: true | false

Default: true

Required: No

tdDomainName :: Lens' TransferDomain Text

The name of a domain.

Type: String

Default: None

Constraints: The domain name can contain only the letters a through z, the numbers 0 through 9, and hyphen (-). Internationalized Domain Names are not supported.

Required: Yes

tdDurationInYears :: Lens' TransferDomain Natural

The number of years the domain will be registered. Domains are registered for a minimum of one year. The maximum period depends on the top-level domain.

Type: Integer

Default: 1

Valid values: Integer from 1 to 10

Required: Yes

tdIdnLangCode :: Lens' TransferDomain (Maybe Text)

Reserved for future use.

tdNameservers :: Lens' TransferDomain [Nameserver]

Contains details for the host and glue IP addresses.

Type: Complex

Children: GlueIps, Name

tdPrivacyProtectAdminContact :: Lens' TransferDomain (Maybe Bool)

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter.

Type: Boolean

Default: true

Valid values: true | false

Required: No

tdPrivacyProtectRegistrantContact :: Lens' TransferDomain (Maybe Bool)

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter.

Type: Boolean

Default: true

Valid values: true | false

Required: No

tdPrivacyProtectTechContact :: Lens' TransferDomain (Maybe Bool)

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter.

Type: Boolean

Default: true

Valid values: true | false

Required: No

tdRegistrantContact :: Lens' TransferDomain ContactDetail

Provides detailed contact information.

Type: Complex

Children: FirstName, MiddleName, LastName, ContactType, OrganizationName, AddressLine1, AddressLine2, City, State, CountryCode, ZipCode, PhoneNumber, Email, Fax, ExtraParams

Required: Yes

tdTechContact :: Lens' TransferDomain ContactDetail

Provides detailed contact information.

Type: Complex

Children: FirstName, MiddleName, LastName, ContactType, OrganizationName, AddressLine1, AddressLine2, City, State, CountryCode, ZipCode, PhoneNumber, Email, Fax, ExtraParams

Required: Yes

Response

data TransferDomainResponse

Response constructor

transferDomainResponse

Arguments

:: Text
tdrOperationId

-> TransferDomainResponse

TransferDomainResponse constructor.

The fields accessible through corresponding lenses are:

tdrOperationId :: Text

Response lenses

tdrOperationId :: Lens' TransferDomainResponse Text

Identifier for tracking the progress of the request. To use this ID to query the operation status, use GetOperationDetail.

Type: String

Default: None

Constraints: Maximum 255 characters.