Network.AWS.Route53Domains.Types

Description

Service Configuration

route53Domains :: Service

API version 2014-05-15 of the Amazon Route 53 Domains SDK configuration.

Errors

_InvalidInput :: AsError a => Getting (First ServiceError) a ServiceError

The requested item is not acceptable. For example, for an OperationId it may refer to the ID of an operation that is already completed. For a domain name, it may not be a valid domain name or belong to the requester account.

_OperationLimitExceeded :: AsError a => Getting (First ServiceError) a ServiceError

The number of operations or jobs running exceeded the allowed threshold for the account.

_DomainLimitExceeded :: AsError a => Getting (First ServiceError) a ServiceError

The number of domains has exceeded the allowed threshold for the account.

_UnsupportedTLD :: AsError a => Getting (First ServiceError) a ServiceError

Amazon Route 53 does not support this top-level domain.

_TLDRulesViolation :: AsError a => Getting (First ServiceError) a ServiceError

The top-level domain does not support this operation.

_DuplicateRequest :: AsError a => Getting (First ServiceError) a ServiceError

The request is already in progress for the domain.

ContactType

data ContactType

Constructors

Association

Company

Person

PublicBody

Reseller

CountryCode

data CountryCode

Constructors

AD

AE

AF

AG

AI

AL

AM

AN

AO

AQ

AR

AS

AT

AU

AW

AZ

BA

BB

BD

BE

BF

BG

BH

BI

BJ

BL

BM

BN

BO

BR

BS

BT

BW

BY

BZ

CA

CC

CD

CF

CG

CH

CI

CK

CL

CM

CN

CO

CR

CU

CV

CX

CY

CZ

DE

DJ

DK

DM

DO

DZ

EC

EE

EG

ER

ES

ET

FI

FJ

FK

FM

FO

FR

GA

GB

GD

GE

GH

GI

GL

GM

GN

GQ

GR

GT'

GU

GW

GY

HK

HN

HR

HT

HU

IE

IL

IM

IN

IQ

IR

IS

IT

Id

JM

JO

JP

KE

KG

KH

KI

KM

KN

KP

KR

KW

KY

KZ

LA

LB

LC

LI

LK

LR

LS

LT'

LU

LV

LY

MA

MC

MD

ME

MF

MG

MH

MK

ML

MM

MN

MO

MP

MR

MS

MT

MU

MV

MW

MX

MY

MZ

NA

NC

NE

NG

NI

NL

NO

NP

NR

NU

NZ

OM

PA

PE

PF

PG

PH

PK

PL

PM

PN

PR

PT

PW

PY

QA

RO

RS

RU

RW

SA

SB

SC

SD

SE

SG

SH

SI

SK

SL

SM

SN

SO

SR

ST

SV

SY

SZ

TC

TD

TG

TH

TJ

TK

TL

TM

TN

TO

TR

TT

TV

TW

TZ

UA

UG

US

UY

UZ

VA

VC

VE

VG

VI

VN

VU

WF

WS

YE

YT

ZA

ZM

ZW

DomainAvailability

data DomainAvailability

Constructors

Available

AvailablePreorder

AvailableReserved

DontKnow

Reserved

Unavailable

UnavailablePremium

UnavailableRestricted

ExtraParamName

data ExtraParamName

Constructors

AuIdNumber

AuIdType

BirthCity

BirthCountry

BirthDateInYyyyMmDd

BirthDepartment

BrandNumber

CaBusinessEntityType

CaLegalType

DocumentNumber

DunsNumber

EsIdentification

EsIdentificationType

EsLegalForm

FiBusinessNumber

FiIdNumber

ItPin

RuPassportData

SeIdNumber

SgIdNumber

VatNumber

OperationStatus

data OperationStatus

Constructors

Error'

Failed

InProgress

Submitted

Successful

OperationType

data OperationType

Constructors

ChangePrivacyProtection

DeleteDomain

DomainLock

RegisterDomain

TransferInDomain

UpdateDomainContact

UpdateNameserver

ReachabilityStatus

data ReachabilityStatus

Constructors

Done

Expired

Pending

BillingRecord

data BillingRecord

Information for one billing record.

See: billingRecord smart constructor.

billingRecord :: BillingRecord

Creates a value of BillingRecord with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

brOperation - The operation that you were charged for.

brInvoiceId - The ID of the invoice that is associated with the billing record.

brDomainName - The name of the domain that the billing record applies to. If the domain name contains characters other than a-z, 0-9, and - (hyphen), such as an internationalized domain name, then this value is in Punycode. For more information, see DNS Domain Name Format in the Amazon Route 53 Developer Guidezzz .

brBillDate - The date that the operation was billed, in Unix format.

brPrice - The price that you were charged for the operation, in US dollars. Example value: 12.0

brOperation :: Lens' BillingRecord (Maybe OperationType)

The operation that you were charged for.

brInvoiceId :: Lens' BillingRecord (Maybe Text)

The ID of the invoice that is associated with the billing record.

brDomainName :: Lens' BillingRecord (Maybe Text)

The name of the domain that the billing record applies to. If the domain name contains characters other than a-z, 0-9, and - (hyphen), such as an internationalized domain name, then this value is in Punycode. For more information, see DNS Domain Name Format in the Amazon Route 53 Developer Guidezzz .

brBillDate :: Lens' BillingRecord (Maybe UTCTime)

The date that the operation was billed, in Unix format.

brPrice :: Lens' BillingRecord (Maybe Double)

The price that you were charged for the operation, in US dollars. Example value: 12.0

ContactDetail

data ContactDetail

ContactDetail includes the following elements.

See: contactDetail smart constructor.

contactDetail :: ContactDetail

Creates a value of ContactDetail with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdOrganizationName - Name of the organization for contact types other than PERSON .

cdEmail - Email address of the contact.

cdState - The state or province of the contact's city.

cdFax - Fax number of the contact. Constraints: Phone number must be specified in the format "+[country dialing code].[number including any area code]". For example, a US phone number might appear as "+1.1234567890" .

cdLastName - Last name of contact.

cdExtraParams - A list of name-value pairs for parameters required by certain top-level domains.

cdZipCode - The zip or postal code of the contact's address.

cdAddressLine1 - First line of the contact's address.

cdCity - The city of the contact's address.

cdPhoneNumber - The phone number of the contact. Constraints: Phone number must be specified in the format "+[country dialing code].[number including any area code>]". For example, a US phone number might appear as "+1.1234567890" .

cdAddressLine2 - Second line of contact's address, if any.

cdFirstName - First name of contact.

cdCountryCode - Code for the country of the contact's address.

cdContactType - Indicates whether the contact is a person, company, association, or public organization. If you choose an option other than PERSON , you must enter an organization name, and you can't enable privacy protection for the contact.

cdOrganizationName :: Lens' ContactDetail (Maybe Text)

Name of the organization for contact types other than PERSON .

cdEmail :: Lens' ContactDetail (Maybe Text)

Email address of the contact.

cdState :: Lens' ContactDetail (Maybe Text)

The state or province of the contact's city.

cdFax :: Lens' ContactDetail (Maybe Text)

Fax number of the contact. Constraints: Phone number must be specified in the format "+[country dialing code].[number including any area code]". For example, a US phone number might appear as "+1.1234567890" .

cdLastName :: Lens' ContactDetail (Maybe Text)

Last name of contact.

cdExtraParams :: Lens' ContactDetail [ExtraParam]

A list of name-value pairs for parameters required by certain top-level domains.

cdZipCode :: Lens' ContactDetail (Maybe Text)

The zip or postal code of the contact's address.

cdAddressLine1 :: Lens' ContactDetail (Maybe Text)

First line of the contact's address.

cdCity :: Lens' ContactDetail (Maybe Text)

The city of the contact's address.

cdPhoneNumber :: Lens' ContactDetail (Maybe Text)

The phone number of the contact. Constraints: Phone number must be specified in the format "+[country dialing code].[number including any area code>]". For example, a US phone number might appear as "+1.1234567890" .

cdAddressLine2 :: Lens' ContactDetail (Maybe Text)

Second line of contact's address, if any.

cdFirstName :: Lens' ContactDetail (Maybe Text)

First name of contact.

cdCountryCode :: Lens' ContactDetail (Maybe CountryCode)

Code for the country of the contact's address.

cdContactType :: Lens' ContactDetail (Maybe ContactType)

Indicates whether the contact is a person, company, association, or public organization. If you choose an option other than PERSON , you must enter an organization name, and you can't enable privacy protection for the contact.

DomainSuggestion

data DomainSuggestion

Information about one suggested domain name.

See: domainSuggestion smart constructor.

domainSuggestion :: DomainSuggestion

Creates a value of DomainSuggestion with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dAvailability - Whether the domain name is available for registering. Valid values: * AVAILABLE * The domain name is available. * AVAILABLE_RESERVED * The domain name is reserved under specific conditions. * AVAILABLE_PREORDER * The domain name is available and can be preordered. * DONT_KNOW * The TLD registry didn't reply with a definitive answer about whether the domain name is available. Amazon Route 53 can return this response for a variety of reasons, for example, the registry is performing maintenance. Try again later. * PENDING * The TLD registry didn't return a response in the expected amount of time. When the response is delayed, it usually takes just a few extra seconds. You can resubmit the request immediately. * RESERVED * The domain name has been reserved for another person or organization. * UNAVAILABLE * The domain name is not available. * UNAVAILABLE_PREMIUM * The domain name is not available. * UNAVAILABLE_RESTRICTED * The domain name is forbidden.

dDomainName - A suggested domain name.

dAvailability :: Lens' DomainSuggestion (Maybe Text)

Whether the domain name is available for registering. Valid values: * AVAILABLE * The domain name is available. * AVAILABLE_RESERVED * The domain name is reserved under specific conditions. * AVAILABLE_PREORDER * The domain name is available and can be preordered. * DONT_KNOW * The TLD registry didn't reply with a definitive answer about whether the domain name is available. Amazon Route 53 can return this response for a variety of reasons, for example, the registry is performing maintenance. Try again later. * PENDING * The TLD registry didn't return a response in the expected amount of time. When the response is delayed, it usually takes just a few extra seconds. You can resubmit the request immediately. * RESERVED * The domain name has been reserved for another person or organization. * UNAVAILABLE * The domain name is not available. * UNAVAILABLE_PREMIUM * The domain name is not available. * UNAVAILABLE_RESTRICTED * The domain name is forbidden.

dDomainName :: Lens' DomainSuggestion (Maybe Text)

A suggested domain name.

DomainSummary

data DomainSummary

Summary information about one domain.

See: domainSummary smart constructor.

domainSummary

Arguments

:: Text
dsDomainName

-> DomainSummary

Creates a value of DomainSummary with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsExpiry - Expiration date of the domain in Coordinated Universal Time (UTC).

dsTransferLock - Indicates whether a domain is locked from unauthorized transfer to another party.

dsAutoRenew - Indicates whether the domain is automatically renewed upon expiration.

dsDomainName - The name of the domain that the summary information applies to.

dsExpiry :: Lens' DomainSummary (Maybe UTCTime)

Expiration date of the domain in Coordinated Universal Time (UTC).

dsTransferLock :: Lens' DomainSummary (Maybe Bool)

Indicates whether a domain is locked from unauthorized transfer to another party.

dsAutoRenew :: Lens' DomainSummary (Maybe Bool)

Indicates whether the domain is automatically renewed upon expiration.

dsDomainName :: Lens' DomainSummary Text

The name of the domain that the summary information applies to.

ExtraParam

data ExtraParam

ExtraParam includes the following elements.

See: extraParam smart constructor.

extraParam

Arguments

:: ExtraParamName
epName

-> Text
epValue

-> ExtraParam

Creates a value of ExtraParam with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

epName - Name of the additional parameter required by the top-level domain.

epValue - Values corresponding to the additional parameter names required by some top-level domains.

epName :: Lens' ExtraParam ExtraParamName

Name of the additional parameter required by the top-level domain.

epValue :: Lens' ExtraParam Text

Values corresponding to the additional parameter names required by some top-level domains.

Nameserver

data Nameserver

Nameserver includes the following elements.

See: nameserver smart constructor.

nameserver

Arguments

:: Text
nName

-> Nameserver

Creates a value of Nameserver with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

nGlueIPs - Glue IP address of a name server entry. Glue IP addresses are required only when the name of the name server is a subdomain of the domain. For example, if your domain is example.com and the name server for the domain is ns.example.com, you need to specify the IP address for ns.example.com. Constraints: The list can contain only one IPv4 and one IPv6 address.

nName - The fully qualified host name of the name server. Constraint: Maximum 255 characters

nGlueIPs :: Lens' Nameserver [Text]

Glue IP address of a name server entry. Glue IP addresses are required only when the name of the name server is a subdomain of the domain. For example, if your domain is example.com and the name server for the domain is ns.example.com, you need to specify the IP address for ns.example.com. Constraints: The list can contain only one IPv4 and one IPv6 address.

nName :: Lens' Nameserver Text

The fully qualified host name of the name server. Constraint: Maximum 255 characters

OperationSummary

data OperationSummary

OperationSummary includes the following elements.

See: operationSummary smart constructor.

operationSummary

Arguments

:: Text
osOperationId

-> OperationStatus
osStatus

-> OperationType
osType

-> UTCTime
osSubmittedDate

-> OperationSummary

Creates a value of OperationSummary with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

osOperationId - Identifier returned to track the requested action.

osStatus - The current status of the requested operation in the system.

osType - Type of the action requested.

osSubmittedDate - The date when the request was submitted.

osOperationId :: Lens' OperationSummary Text

Identifier returned to track the requested action.

osStatus :: Lens' OperationSummary OperationStatus

The current status of the requested operation in the system.

osType :: Lens' OperationSummary OperationType

Type of the action requested.

osSubmittedDate :: Lens' OperationSummary UTCTime

The date when the request was submitted.

Tag

data Tag

Each tag includes the following elements.

See: tag smart constructor.

tag :: Tag

Creates a value of Tag with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tagValue - The value of a tag. Valid values: A-Z, a-z, 0-9, space, ".:/=+-@" Constraints: Each value can be 0-256 characters long.

tagKey - The key (name) of a tag. Valid values: A-Z, a-z, 0-9, space, ".:/=+-@" Constraints: Each key can be 1-128 characters long.

tagValue :: Lens' Tag (Maybe Text)

The value of a tag. Valid values: A-Z, a-z, 0-9, space, ".:/=+-@" Constraints: Each value can be 0-256 characters long.

tagKey :: Lens' Tag (Maybe Text)

The key (name) of a tag. Valid values: A-Z, a-z, 0-9, space, ".:/=+-@" Constraints: Each key can be 1-128 characters long.