Network.AWS.Route53Domains.UpdateDomainContactPrivacy

Description

This operation updates the specified domain contact's privacy setting. When the privacy option is enabled, personal information such as postal or email address is hidden from the results of a public WHOIS query. The privacy services are provided by the AWS registrar, Gandi. For more information, see the Gandi privacy features .

This operation only affects the privacy of the specified contact type (registrant, administrator, or tech). Successful acceptance returns an operation ID that you can use with GetOperationDetail to track the progress and completion of the action. If the request is not completed successfully, the domain registrant will be notified by email.

Creating a Request

updateDomainContactPrivacy

Arguments

:: Text
udcpDomainName

-> UpdateDomainContactPrivacy

Creates a value of UpdateDomainContactPrivacy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

udcpTechPrivacy - Whether you want to conceal contact information from WHOIS queries. If you specify true , WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter.

udcpRegistrantPrivacy - Whether you want to conceal contact information from WHOIS queries. If you specify true , WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter.

udcpAdminPrivacy - Whether you want to conceal contact information from WHOIS queries. If you specify true , WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter.

udcpDomainName - The name of the domain that you want to update the privacy setting for.

data UpdateDomainContactPrivacy

The UpdateDomainContactPrivacy request includes the following elements.

See: updateDomainContactPrivacy smart constructor.

Request Lenses

udcpTechPrivacy :: Lens' UpdateDomainContactPrivacy (Maybe Bool)

Whether you want to conceal contact information from WHOIS queries. If you specify true , WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter.

udcpRegistrantPrivacy :: Lens' UpdateDomainContactPrivacy (Maybe Bool)

Whether you want to conceal contact information from WHOIS queries. If you specify true , WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter.

udcpAdminPrivacy :: Lens' UpdateDomainContactPrivacy (Maybe Bool)

Whether you want to conceal contact information from WHOIS queries. If you specify true , WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter.

udcpDomainName :: Lens' UpdateDomainContactPrivacy Text

The name of the domain that you want to update the privacy setting for.

Destructuring the Response

updateDomainContactPrivacyResponse

Arguments

:: Int
udcprsResponseStatus

-> Text
udcprsOperationId

-> UpdateDomainContactPrivacyResponse

Creates a value of UpdateDomainContactPrivacyResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

udcprsResponseStatus - -- | The response status code.

udcprsOperationId - Identifier for tracking the progress of the request. To use this ID to query the operation status, use GetOperationDetail.

data UpdateDomainContactPrivacyResponse

The UpdateDomainContactPrivacy response includes the following element.

See: updateDomainContactPrivacyResponse smart constructor.

Response Lenses

udcprsResponseStatus :: Lens' UpdateDomainContactPrivacyResponse Int

- | The response status code.

udcprsOperationId :: Lens' UpdateDomainContactPrivacyResponse Text

Identifier for tracking the progress of the request. To use this ID to query the operation status, use GetOperationDetail.