Network.AWS.Route53Domains.UpdateDomainNameservers

Description

This operation replaces the current set of name servers for the domain with the specified set of name servers. If you use Amazon Route 53 as your DNS service, specify the four name servers in the delegation set for the hosted zone for the domain.

If successful, this operation returns an operation ID that you can use to track the progress and completion of the action. If the request is not completed successfully, the domain registrant will be notified by email.

Creating a Request

updateDomainNameservers

Arguments

:: Text
udnDomainName

-> UpdateDomainNameservers

Creates a value of UpdateDomainNameservers with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

udnFIAuthKey - The authorization key for .fi domains

udnDomainName - The name of the domain that you want to change name servers for.

udnNameservers - A list of new name servers for the domain.

data UpdateDomainNameservers

Replaces the current set of name servers for the domain with the specified set of name servers. If you use Amazon Route 53 as your DNS service, specify the four name servers in the delegation set for the hosted zone for the domain.

If successful, this operation returns an operation ID that you can use to track the progress and completion of the action. If the request is not completed successfully, the domain registrant will be notified by email.

See: updateDomainNameservers smart constructor.

Request Lenses

udnFIAuthKey :: Lens' UpdateDomainNameservers (Maybe Text)

The authorization key for .fi domains

udnDomainName :: Lens' UpdateDomainNameservers Text

The name of the domain that you want to change name servers for.

udnNameservers :: Lens' UpdateDomainNameservers [Nameserver]

A list of new name servers for the domain.

Destructuring the Response

updateDomainNameserversResponse

Arguments

:: Int
udnrsResponseStatus

-> Text
udnrsOperationId

-> UpdateDomainNameserversResponse

Creates a value of UpdateDomainNameserversResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

udnrsResponseStatus - -- | The response status code.

udnrsOperationId - Identifier for tracking the progress of the request. To use this ID to query the operation status, use GetOperationDetail .

data UpdateDomainNameserversResponse

The UpdateDomainNameservers response includes the following element.

See: updateDomainNameserversResponse smart constructor.

Response Lenses

udnrsResponseStatus :: Lens' UpdateDomainNameserversResponse Int

- | The response status code.

udnrsOperationId :: Lens' UpdateDomainNameserversResponse Text

Identifier for tracking the progress of the request. To use this ID to query the operation status, use GetOperationDetail .