Network.AWS.Route53Domains.UpdateDomainNameservers

Description

This operation replaces the current set of name servers for the domain with the specified set of name servers. If you use Amazon Route 53 as your DNS service, specify the four name servers in the delegation set for the hosted zone for the domain.

If successful, this operation returns an operation ID that you can use to track the progress and completion of the action. If the request is not completed successfully, the domain registrant will be notified by email.

http://docs.aws.amazon.com/Route53/latest/APIReference/api-UpdateDomainNameservers.html

Request

data UpdateDomainNameservers

Request constructor

updateDomainNameservers

Arguments

:: Text
udnDomainName

-> UpdateDomainNameservers

UpdateDomainNameservers constructor.

The fields accessible through corresponding lenses are:

udnDomainName :: Text

udnFIAuthKey :: Maybe Text

udnNameservers :: [Nameserver]

Request lenses

udnDomainName :: Lens' UpdateDomainNameservers Text

The name of a domain.

Type: String

Default: None

Constraints: The domain name can contain only the letters a through z, the numbers 0 through 9, and hyphen (-). Internationalized Domain Names are not supported.

Required: Yes

udnFIAuthKey :: Lens' UpdateDomainNameservers (Maybe Text)

The authorization key for .fi domains

udnNameservers :: Lens' UpdateDomainNameservers [Nameserver]

A list of new name servers for the domain.

Type: Complex

Children: Name, GlueIps

Required: Yes

Response

data UpdateDomainNameserversResponse

Response constructor

updateDomainNameserversResponse

Arguments

:: Text
udnrOperationId

-> UpdateDomainNameserversResponse

UpdateDomainNameserversResponse constructor.

The fields accessible through corresponding lenses are:

udnrOperationId :: Text

Response lenses

udnrOperationId :: Lens' UpdateDomainNameserversResponse Text

Identifier for tracking the progress of the request. To use this ID to query the operation status, use GetOperationDetail.

Type: String

Default: None

Constraints: Maximum 255 characters.