Network.AWS.Route53Domains.V2014_05_15.CheckDomainAvailability

Description

This operation checks the availability of one domain name. You can access this API without authenticating. Note that if the availability status of a domain is pending, you must submit another request to determine the availability of the domain name. CheckDomainAvailability Example POST / HTTP/1.1 host:route53domains.us-east-1.amazonaws.com x-amz-date:20140711T205225Z authorization:AWS4-HMAC-SHA256 Credential=AKIAIOSFODNN7EXAMPLE20140711us-east-1route53domainsaws4_request,

SignedHeaders=content-length;content-type;host;user-agent;x-amz-date;x-amz-target, Signature=[calculated-signature] x-amz-target:Route53Domains_v20140515.CheckDomainAvailability user-agent:aws-sdk-java1.8.3 Linux2.6.18-164.el5PAE Java_HotSpot (TM )_Server_VM24.60-b091.7.0_60 content-type:application/x-amz-json- 1.1 content-length:[number of characters in the JSON string] connections:Keep-Alive { DomainName:"example.com" } HTTP/1.1 200 Content-Length:[number of characters in the JSON string] { Availability:AVAILABLE }.

Request

data CheckDomainAvailability

The CheckDomainAvailability request contains the following elements.

Request constructor

mkCheckDomainAvailability

Arguments

:: Text
cdaDomainName

-> CheckDomainAvailability

Smart constructor for the minimum required parameters to construct a valid CheckDomainAvailability request.

Request lenses

cdaDomainName :: Lens' CheckDomainAvailability Text

The name of a domain. Type: String Default: None Constraints: The domain name can contain only the letters a through z, the numbers 0 through 9, and hyphen (-). Internationalized Domain Names are not supported. Required: Yes.

cdaIdnLangCode :: Lens' CheckDomainAvailability (Maybe Text)

Reserved for future use.

Response

data CheckDomainAvailabilityResponse

The CheckDomainAvailability response includes the following elements.

Response lenses

cdarAvailability :: Lens' CheckDomainAvailabilityResponse DomainAvailability

Whether the domain name is available for registering. You can only register domains designated as AVAILABLE. Type: String Valid values: AVAILABLE – The domain name is available. AVAILABLE_RESERVED – The domain name is reserved under specific conditions. AVAILABLE_PREORDER – The domain name is available and can be preordered. UNAVAILABLE – The domain name is not available. UNAVAILABLE_PREMIUM – The domain name is not available. UNAVAILABLE_RESTRICTED – The domain name is forbidden. RESERVED – The domain name has been reserved for another person or organization.