Network.AWS.Route53Domains.V2014_05_15.GetDomainDetail

Description

This operation returns detailed information about the domain. The domain's contact information is also returned as part of the output. GetDomainDetail Example POST HTTP1.1 host:route53domains.us-east-1.amazonaws.com x-amz-date:20140711T205230Z authorization:AWS4-HMAC-SHA256 Credential=AKIAIOSFODNN7EXAMPLE20140711us-east-1route53domainsaws4_request,

SignedHeaders=content-length;content-type;host;user-agent;x-amz-date;x-amz-target, Signature=[calculated-signature] x-amz-target:Route53Domains_v20140515.GetDomainDetail user-agent:aws-sdk-java1.8.3 Linux2.6.18-164.el5PAE Java_HotSpot (TM )_Server_VM24.60-b091.7.0_60 content-type:application/x-amz-json-1.1 content-length:[number of characters in the JSON string] { DomainName:"example.com" } HTTP/1.1 200 Content-Length:[number of characters in the JSON string] { AbuseContactEmail:"abusesupport.gandi.net", AbuseContactPhone:"+33.170377661", AdminContact:{ AddressLine1:"1 Any Street", AddressLine2:"", City:Anytown, CountryCode:US, Email:"johnexample.com", ExtraParams:[ ], FirstName:John, LastName:Doe, PhoneNumber:"+2065550100", State:WA, ZipCode:"98101" }, AdminPrivacy:true, AutoRenew:true, CreationDate:1400010459, DomainName:"example.com", ExpirationDate:1431539259, Nameservers:[ { GlueIps:[ ], Name:"ns-2048.awsdns-64.com" }, { GlueIps:[ ], Name:"ns-2051.awsdns-67.co.uk" }, { GlueIps:[ ], Name:"ns-2050.awsdns-66.org" }, { GlueIps:[ ], Name:"ns-2049.awsdns-65.net" } ], RegistrantContact:{ AddressLine1:"1 Any Street", AddressLine2:"", City:Anytown, CountryCode:US, Email:"johnexample.com", ExtraParams:[ ], FirstName:John, LastName:Doe, PhoneNumber:"+2065550100", State:WA, ZipCode:"98101" }, RegistrantPrivacy:true, RegistrarName:"GANDI SAS", RegistrarUrl:"http://www.gandi.net", Reseller:Amazon, StatusList:[ "clientTransferProhibited" ], TechContact:{ AddressLine1:"1 Any Street", AddressLine2:"", City:Anytown, CountryCode:US, Email:"johnexample.com", ExtraParams:[ ], FirstName:John, LastName:Doe, PhoneNumber:"+2065550100", State:WA, ZipCode:"98101" }, TechPrivacy:true, UpdatedDate:1400010459, WhoIsServer:"whois.gandi.net" }.

Documentation

data GetDomainDetail

Constructors

GetDomainDetail

Fields

_gddrDomainName :: Text

The name of a domain. Type: String Default: None Constraints: The domain name can contain only the letters a through z, the numbers 0 through 9, and hyphen (-). Internationalized Domain Names are not supported. Required: Yes.

gddrDomainName :: Iso' GetDomainDetail Text

data GetDomainDetailResponse

Constructors

GetDomainDetailResponse

Fields

_gddsRegistrantContact :: ContactDetail

Provides details about the domain registrant. Type: Complex Children: FirstName, MiddleName, LastName, ContactType, OrganizationName, AddressLine1, AddressLine2, City, State, CountryCode, ZipCode, PhoneNumber, Email, Fax, ExtraParams.

_gddsAdminContact :: ContactDetail

Provides details about the domain administrative contact. Type: Complex Children: FirstName, MiddleName, LastName, ContactType, OrganizationName, AddressLine1, AddressLine2, City, State, CountryCode, ZipCode, PhoneNumber, Email, Fax, ExtraParams.

_gddsTechContact :: ContactDetail

Provides details about the domain technical contact. Type: Complex Children: FirstName, MiddleName, LastName, ContactType, OrganizationName, AddressLine1, AddressLine2, City, State, CountryCode, ZipCode, PhoneNumber, Email, Fax, ExtraParams.

_gddsDomainName :: Text

The name of a domain. Type: String.

_gddsNameservers :: [Nameserver]

The name of the domain. Type: String.

_gddsTechPrivacy :: Maybe Bool

Specifies whether contact information for the tech contact is concealed from WHOIS queries. If the value is true, WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter. Type: Boolean.

_gddsRegistrantPrivacy :: Maybe Bool

Specifies whether contact information for the registrant contact is concealed from WHOIS queries. If the value is true, WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter. Type: Boolean.

_gddsAdminPrivacy :: Maybe Bool

Specifies whether contact information for the admin contact is concealed from WHOIS queries. If the value is true, WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter. Type: Boolean.

_gddsAutoRenew :: Maybe Bool

Specifies whether the domain registration is set to renew automatically. Type: Boolean.

_gddsAbuseContactPhone :: Maybe Text

Phone number for reporting abuse. Type: String.

_gddsDnsSec :: Maybe Text

Reserved for future use.

_gddsStatusList :: [Text]

An array of domain name status codes, also known as Extensible Provisioning Protocol (EPP) status codes. ICANN, the organization that maintains a central database of domain names, has developed a set of domain name status codes that tell you the status of a variety of operations on a domain name, for example, registering a domain name, transferring a domain name to another registrar, renewing the registration for a domain name, and so on. All registrars use this same set of status codes. For a current list of domain name status codes and an explanation of what each code means, go to the ICANN website and search for epp status codes. (Search on the ICANN website; web searches sometimes return an old version of the document.) Type: Array of String.

_gddsAbuseContactEmail :: Maybe Text

Email address to contact to report incorrect contact information for a domain, to report that the domain is being used to send spam, to report that someone is cybersquatting on a domain name, or report some other type of abuse. Type: String.

_gddsRegistrarName :: Maybe Text

Name of the registrar of the domain as identified in the registry. Amazon Route 53 domains are registered by registrar Gandi. The value is "GANDI SAS". Type: String.

_gddsRegistrarUrl :: Maybe Text

Web address of the registrar. Type: String.

_gddsWhoIsServer :: Maybe Text

The fully qualified name of the WHOIS server that can answer the WHOIS query for the domain. Type: String.

_gddsRegistryDomainId :: Maybe Text

Reserved for future use.

_gddsReseller :: Maybe Text

Reseller of the domain. Domains registered or transferred using Amazon Route 53 domains will have Amazon as the reseller. Type: String.

_gddsUpdatedDate :: Maybe ISO8601

The last updated date of the domain as found in the response to a WHOIS query. The date format is Unix time.

_gddsExpirationDate :: Maybe ISO8601

The date when the registration for the domain is set to expire. The date format is Unix time.

_gddsCreationDate :: Maybe ISO8601

The date when the domain was created as found in the response to a WHOIS query. The date format is Unix time.

gddsWhoIsServer :: Lens' GetDomainDetailResponse (Maybe Text)

gddsUpdatedDate :: Lens' GetDomainDetailResponse (Maybe ISO8601)

gddsTechPrivacy :: Lens' GetDomainDetailResponse (Maybe Bool)

gddsTechContact :: Lens' GetDomainDetailResponse ContactDetail

gddsStatusList :: Lens' GetDomainDetailResponse [Text]

gddsReseller :: Lens' GetDomainDetailResponse (Maybe Text)

gddsRegistryDomainId :: Lens' GetDomainDetailResponse (Maybe Text)

gddsRegistrarUrl :: Lens' GetDomainDetailResponse (Maybe Text)

gddsRegistrarName :: Lens' GetDomainDetailResponse (Maybe Text)

gddsRegistrantPrivacy :: Lens' GetDomainDetailResponse (Maybe Bool)

gddsRegistrantContact :: Lens' GetDomainDetailResponse ContactDetail

gddsNameservers :: Lens' GetDomainDetailResponse [Nameserver]

gddsExpirationDate :: Lens' GetDomainDetailResponse (Maybe ISO8601)

gddsDomainName :: Lens' GetDomainDetailResponse Text

gddsDnsSec :: Lens' GetDomainDetailResponse (Maybe Text)

gddsCreationDate :: Lens' GetDomainDetailResponse (Maybe ISO8601)

gddsAutoRenew :: Lens' GetDomainDetailResponse (Maybe Bool)

gddsAdminPrivacy :: Lens' GetDomainDetailResponse (Maybe Bool)

gddsAdminContact :: Lens' GetDomainDetailResponse ContactDetail

gddsAbuseContactPhone :: Lens' GetDomainDetailResponse (Maybe Text)

gddsAbuseContactEmail :: Lens' GetDomainDetailResponse (Maybe Text)